refactor: split gateway main runtime assembly
