+++ cmd/gateway/main.go
@@ -7,142 +7,25 @@ import (
 	"expvar"
 	"flag"
 	"fmt"
-	"io"
 	"log"
-	"net"
 	"net/http"
 	"os"
 	"os/signal"
-	"regexp"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"syscall"
 	"time"
 
-	"github.com/Project-Helianthus/helianthus-ebusgateway"
+	ebusgateway "github.com/Project-Helianthus/helianthus-ebusgateway"
 	"github.com/Project-Helianthus/helianthus-ebusgateway/graphql"
-	"github.com/Project-Helianthus/helianthus-ebusgateway/internal/adaptermux"
 	"github.com/Project-Helianthus/helianthus-ebusgateway/internal/adaptermux/v8classifier"
 	"github.com/Project-Helianthus/helianthus-ebusgateway/internal/eebusadmin"
 	"github.com/Project-Helianthus/helianthus-ebusgateway/internal/runtimestate"
 	"github.com/Project-Helianthus/helianthus-ebusgateway/mcp"
-	"github.com/Project-Helianthus/helianthus-ebusgateway/mcp/ebus_standard"
 	"github.com/Project-Helianthus/helianthus-ebusgateway/mdns"
-	"github.com/Project-Helianthus/helianthus-ebusgateway/portal"
-	"github.com/Project-Helianthus/helianthus-ebusgateway/ui"
-	ebuserrors "github.com/Project-Helianthus/helianthus-ebusgo/errors"
 	"github.com/Project-Helianthus/helianthus-ebusgo/protocol"
-	ebusgoTransport "github.com/Project-Helianthus/helianthus-ebusgo/transport"
 	vaillantproviders "github.com/Project-Helianthus/helianthus-ebusreg/providers/vaillant"
-	"github.com/Project-Helianthus/helianthus-ebusreg/registry"
-	"github.com/Project-Helianthus/helianthus-ebusreg/vaillant/productids"
 )
-
-var (
-	buildVersion                              = "dev"
-	buildID                                   = "unknown"
-	wireObserveFirstObserversFn               = wireObserveFirstObservers
-	startDiscoveryScanLoopFn                  = startDiscoveryScanLoopWithClassifier
-	startVaillantSemanticPollingFn            = startVaillantSemanticPolling
-	attachPassiveShadowProducerFn             = (*vaillantSemanticPoller).AttachPassiveShadowProducer
-	startPassiveTransactionReconstructor      = ebusgateway.StartPassiveTransactionReconstructor
-	startBroadcastListenerWithReconstructorFn = ebusgateway.StartBroadcastListenerWithReconstructor
-	startHTTPServerFn                         func(context.Context, ebusgateway.Config, *ebusgateway.Gateway, *graphql.Builder, *graphql.BroadcastHub, graphql.SemanticProvider, mcp.ScheduleWriter, mcp.ConfigWriter, *ebusgateway.BusObservabilityStore, *ebusgateway.ShadowCache) (*http.Server, mdns.Advertiser, error)
-	admissionStabilityRefreshDelay            = time.Duration(ebusgateway.StartupAdmissionStateMinStabilitySecondsDefault)*time.Second + 200*time.Millisecond
-	instanceGUIDPattern                       = regexp.MustCompile(`^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$`)
-
-	// v8RolloutExpvarCurrent holds the currently-active bus +
-	// classifier references that back the five helianthus_round9_* /
-	// helianthus_payload_aa_* / helianthus_v8_shadow_* expvars
-	// mirrored onto /debug/vars.
-	//
-	// Why a process-global atomic and not a plain sync.Once-captured
-	// closure: expvar.Publish panics on duplicate names, so the
-	// publishes themselves must be one-shot. But the test harness re-
-	// enters run() with a fresh *Gateway / *protocol.Bus and a fresh
-	// *adaptermux.Mux per scenario. A naive sync.Once that captures
-	// the first bus would leave /debug/vars reading stale counters
-	// from the dead first gateway while /metrics serves the live
-	// counters from the new BusObservabilityStore provider — a
-	// genuine surface-inconsistency bug flagged in PR #655 round-1
-	// review.
-	//
-	// Indirection: publish the expvar.Funcs ONCE with closures that
-	// dereference v8RolloutExpvarCurrent.Load() at every scrape.
-	// run() updates the pointer on each invocation; old gateways
-	// stop being scraped because nothing holds their bus reference
-	// any more.
-	v8RolloutExpvarPublishOnce sync.Once
-	v8RolloutExpvarCurrent     atomic.Pointer[v8RolloutExpvarSource]
-
-	// v8AdminEventsCurrentClassifier is the indirection layer for
-	// the /debug/v8/admin-events HTTP surface. Same rationale as
-	// v8RolloutExpvarCurrent: the handler is registered ONCE at
-	// startHTTPServer and dereferences this atomic.Pointer at
-	// request time. run() updates the pointer on each invocation
-	// (test re-entry safe). When the active transport isn't
-	// adapter-direct, classifier is nil; the handler returns an
-	// empty events array + dropped=0 in that case so the contract
-	// degrades cleanly without 404-ing.
-	v8AdminEventsCurrentClassifier atomic.Pointer[v8classifier.Classifier]
-	v8AdminEventsCurrentProvider   atomic.Pointer[v8AdminClassifierProvider]
-)
-
-// v8RolloutExpvarSource is the indirection layer for the five
-// helianthus_*_total expvar surfaces. Nil-classifier is intentional —
-// non-adapter-direct transports run with classifier=nil and rely on
-// v8classifier.Classifier.ShadowWouldHaveDroppedTotal()'s nil-receiver
-// contract (returns 0). bus must not be nil: gateway.Bus is always
-// non-nil after a successful ebusgateway.New(), and the wiring site
-// guards on it.
-type v8RolloutExpvarSource struct {
-	bus               *protocol.Bus
-	shadowDropCountFn func() uint64
-}
-
-type v8AdminClassifierProvider struct {
-	withClassifier func(func(*v8classifier.Classifier))
-}
-
-func withCurrentV8AdminClassifier(callback func(*v8classifier.Classifier)) {
-	if callback == nil {
-		return
-	}
-	// Direct pointer is retained as a test/backward-compatible injection seam.
-	// Production clears it and installs the generation-aware provider below.
-	if classifier := v8AdminEventsCurrentClassifier.Load(); classifier != nil {
-		callback(classifier)
-		return
-	}
-	provider := v8AdminEventsCurrentProvider.Load()
-	if provider == nil || provider.withClassifier == nil {
-		return
-	}
-	provider.withClassifier(callback)
-}
-
-type runtimeWatchObserver struct {
-	primary  ebusgateway.WatchObserver
-	fallback ebusgateway.WatchObserver
-}
-
-func (observer *runtimeWatchObserver) Observe(key ebusgateway.WatchKey) ebusgateway.WatchObservation {
-	if key == nil {
-		return ebusgateway.WatchObservation{State: ebusgateway.WatchObservationStateCatalogMiss}
-	}
-	if observer != nil && observer.primary != nil {
-		observation := observer.primary.Observe(key)
-		if observation.State != ebusgateway.WatchObservationStateCatalogMiss {
-			return observation
-		}
-	}
-	if observer != nil && observer.fallback != nil {
-		return observer.fallback.Observe(key)
-	}
-	return ebusgateway.WatchObservation{State: ebusgateway.WatchObservationStateCatalogMiss}
-}
 
 func main() {
 	cfg := ebusgateway.DefaultConfig()
@@ -162,28 +45,6 @@ func main() {
 		}
 		log.Fatalf("gateway: %v", err)
 	}
-}
-
-func recordBusAdmissionTransitionWithStabilityRefresh(ctx context.Context, store *ebusgateway.BusObservabilityStore, state string, source, companionTarget byte, reason string) {
-	if store == nil {
-		return
-	}
-	if store.RecordBusAdmissionTransition(state, source, companionTarget, reason) {
-		return
-	}
-	if state != "active" && state != "degraded" {
-		return
-	}
-	go func() {
-		timer := time.NewTimer(admissionStabilityRefreshDelay)
-		defer timer.Stop()
-		select {
-		case <-ctx.Done():
-			return
-		case <-timer.C:
-			store.RecordBusAdmissionTransition(state, source, companionTarget, reason)
-		}
-	}()
 }
 
 func run(ctx context.Context, cfg ebusgateway.Config) (result error) {
@@ -1269,1911 +1130,4 @@ func run(ctx context.Context, cfg ebusgateway.Config) (result error) {
 	}
 	<-ctx.Done()
 	return nil
-}
-
-func shouldStartPassiveObserveFirst(cfg ebusgateway.Config) bool {
-	return cfg.BroadcastListen && ebusgateway.PassiveTransportSupported(cfg)
-}
-
-func waitForStartupScanFirstPass(ctx context.Context, cfg ebusgateway.Config, firstPassDone <-chan struct{}) {
-	if !cfg.ScanOnStart || !shouldStartPassiveObserveFirst(cfg) || firstPassDone == nil {
-		return
-	}
-	select {
-	case <-ctx.Done():
-	case <-firstPassDone:
-	}
-}
-
-func wireObserveFirstObservers(cfg *ebusgateway.Config) (*ebusgateway.BusObservabilityStore, *ebusgateway.ActivePassiveDeduplicator, error) {
-	if cfg == nil {
-		return nil, nil, nil
-	}
-
-	observer := &runtimeWatchObserver{primary: cfg.WatchObserver}
-
-	dedupCfg := *cfg
-	dedupCfg.WatchObserver = cfg.WatchObserver
-
-	var deduplicator *ebusgateway.ActivePassiveDeduplicator
-	if cfg.BroadcastListen {
-		dedup, err := ebusgateway.NewActivePassiveDeduplicator(dedupCfg)
-		if err != nil {
-			return nil, nil, err
-		}
-		deduplicator = dedup
-		observer.fallback = deduplicator
-	}
-
-	observerCfg := *cfg
-	observerCfg.WatchObserver = observer
-	if deduplicator != nil {
-		observerCfg.LocalAddressSnapshotter = deduplicator
-	}
-
-	busObservability := ebusgateway.NewBusObservabilityStore(observerCfg)
-	cfg.BusConfig.Observer = ebusgateway.ChainBusObservers(cfg.BusConfig.Observer, busObservability)
-	cfg.WatchObserver = observer
-	cfg.WatchEfficiencyObserver = busObservability
-	if deduplicator != nil {
-		cfg.BusConfig.Observer = ebusgateway.ChainBusObservers(cfg.BusConfig.Observer, deduplicator)
-	}
-
-	return busObservability, deduplicator, nil
-}
-
-func applyTransportSourcePolicy(cfg *ebusgateway.Config) {
-	if cfg == nil {
-		return
-	}
-	// Resolve a valid URI override once at the process integration boundary so
-	// every downstream admission/source/evidence decision sees the same
-	// canonical transport tuple as the driver factory. Invalid descriptors stay
-	// untouched and are classified driver-locally by DriverManager.
-	if normalized, err := ebusgateway.NormalizeEBusDriverTransportConfig(cfg.TransportConfig); err == nil {
-		cfg.TransportConfig = normalized
-	}
-
-	protocol := strings.TrimSpace(strings.ToLower(string(cfg.TransportConfig.Protocol)))
-	switch protocol {
-	case "ebusd", "ebusd-tcp":
-		return
-	default:
-		if cfg.ScanSourceAuto {
-			cfg.ScanSource = 0x00
-		}
-	}
-}
-
-func admittedMutationSourceForGateway(cfg ebusgateway.Config, admissionPath ebusgateway.TransportAdmissionPath, overrideSet bool) (byte, bool) {
-	if cfg.ScanSource != 0 && !cfg.ScanSourceAuto &&
-		(admissionPath != ebusgateway.TransportAdmissionSourceSelectionCapable || overrideSet) {
-		return cfg.ScanSource, true
-	}
-	if admissionPath == ebusgateway.TransportAdmissionStaticFallback && isEbusdTransportProtocol(cfg.TransportConfig.Protocol) && cfg.ScanSourceAuto && cfg.ScanSource == 0 {
-		defaultSource := ebusgateway.DefaultConfig().ScanSource
-		return defaultSource, defaultSource != 0
-	}
-	return 0, false
-}
-
-func configuredStartupSourceAdmissionCandidate(cfg ebusgateway.Config, admissionPath ebusgateway.TransportAdmissionPath, overrideSet bool) (protocol.SourceAddressSelection, bool) {
-	if admissionPath != ebusgateway.TransportAdmissionSourceSelectionCapable || overrideSet || shouldStartPassiveObserveFirst(cfg) {
-		return protocol.SourceAddressSelection{}, false
-	}
-	if cfg.ScanSourceAuto || cfg.ScanSource == 0x00 {
-		return protocol.SourceAddressSelection{}, false
-	}
-	companion := cfg.StartupCompanionTarget
-	if companion == 0x00 {
-		companion = protocol.CompanionAddressForSource(cfg.ScanSource)
-	}
-	return protocol.SourceAddressSelection{
-		Source:    cfg.ScanSource,
-		Companion: companion,
-		Metrics: protocol.SourceAddressSelectionMetrics{
-			ObservedProbableTargets: append([]byte(nil), cfg.StartupProbeTargets...),
-		},
-	}, true
-}
-
-func isEbusdTransportProtocol(protocol ebusgateway.TransportProtocol) bool {
-	switch strings.TrimSpace(strings.ToLower(string(protocol))) {
-	case "ebusd", string(ebusgateway.TransportEbusdTCP):
-		return true
-	default:
-		return false
-	}
-}
-
-func normalizeInstanceGUID(value string) (string, error) {
-	normalized := strings.TrimSpace(strings.ToLower(value))
-	if normalized == "" {
-		return "", nil
-	}
-	if !instanceGUIDPattern.MatchString(normalized) {
-		return "", fmt.Errorf("invalid instance-guid %q", value)
-	}
-	return normalized, nil
-}
-
-func gatewayMDNSText(cfg ebusgateway.Config) []string {
-	text := []string{
-		"path=" + cfg.GraphQLPath,
-		"transport=http",
-		"version=1",
-	}
-	if cfg.InstanceGUID != "" {
-		text = append(text, "instance_guid="+cfg.InstanceGUID)
-	}
-	return text
-}
-
-func bindFlags(fs *flag.FlagSet, cfg *ebusgateway.Config) *gatewayFlagInputs {
-	inputs := &gatewayFlagInputs{}
-	if fs == nil || cfg == nil {
-		return inputs
-	}
-
-	fs.StringVar((*string)(&cfg.TransportConfig.Protocol), "transport", string(cfg.TransportConfig.Protocol), "transport protocol: enh, ens, udp-plain, tcp-plain, or ebusd-tcp")
-	fs.StringVar(&cfg.TransportConfig.Network, "network", cfg.TransportConfig.Network, "transport network: unix, tcp, or udp")
-	fs.StringVar(&cfg.TransportConfig.Address, "address", cfg.TransportConfig.Address, "transport address (unix socket path or host:port)")
-	fs.DurationVar(&cfg.TransportConfig.ReadTimeout, "read-timeout", cfg.TransportConfig.ReadTimeout, "transport read timeout")
-	fs.DurationVar(&cfg.TransportConfig.WriteTimeout, "write-timeout", cfg.TransportConfig.WriteTimeout, "transport write timeout")
-	fs.DurationVar(&cfg.TransportConfig.DialTimeout, "dial-timeout", cfg.TransportConfig.DialTimeout, "transport dial timeout")
-	fs.BoolVar(&cfg.ModbusTCPConfig.Enabled, "modbus-tcp-enabled", cfg.ModbusTCPConfig.Enabled, "enable the read-only Modbus TCP sidecar")
-	fs.StringVar(&cfg.ModbusTCPConfig.Endpoint, "modbus-tcp-endpoint", cfg.ModbusTCPConfig.Endpoint, "Modbus TCP endpoint URI (tcp://host:port)")
-	fs.StringVar(&inputs.modbusEndpointFile, "modbus-tcp-endpoint-file", "", "path to an owner-only file containing the Modbus TCP endpoint URI")
-	fs.DurationVar(&cfg.ModbusTCPConfig.DialTimeout, "modbus-tcp-dial-timeout", cfg.ModbusTCPConfig.DialTimeout, "Modbus TCP dial timeout")
-	bindEEBusFlags(fs, cfg)
-	bindM2MGraphQLFlags(fs, cfg)
-	fs.BoolVar(
-		&cfg.EvidenceOneShotEnabled,
-		"synchronized-evidence-one-shot-enabled",
-		cfg.EvidenceOneShotEnabled,
-		"enable the fixed owner-only synchronized evidence one-shot control",
-	)
-	fs.IntVar(&cfg.QueueCapacity, "queue-capacity", cfg.QueueCapacity, "bus queue capacity (0 uses protocol default)")
-	fs.BoolVar(&cfg.ScanOnStart, "scan", cfg.ScanOnStart, "scan bus on startup")
-	fs.DurationVar(&cfg.ScanTimeout, "scan-timeout", cfg.ScanTimeout, "startup scan timeout")
-	fs.DurationVar(&cfg.ScanRequestTimeout, "scan-request-timeout", cfg.ScanRequestTimeout, "startup scan per-request timeout")
-	fs.DurationVar(&cfg.ScanInterval, "scan-interval", cfg.ScanInterval, "startup scan retry interval (when scan finds 0 devices)")
-	fs.BoolVar(&cfg.DiagnosticFullRangeRetry, "diagnostic-full-range-retry", cfg.DiagnosticFullRangeRetry, "allow full-range retry on non-ebusd-tcp transports after a Vaillant root candidate is observed")
-	fs.DurationVar(&cfg.BootLiveTimeout, "boot-live-timeout", cfg.BootLiveTimeout, "semantic startup timeout before entering degraded mode")
-	fs.DurationVar(&cfg.SemanticDiscoveryInterval, "semantic-discovery-interval", cfg.SemanticDiscoveryInterval, "semantic discovery polling interval")
-	fs.DurationVar(&cfg.SemanticConfigInterval, "semantic-config-interval", cfg.SemanticConfigInterval, "semantic config polling interval")
-	fs.DurationVar(&cfg.SemanticStateInterval, "semantic-state-interval", cfg.SemanticStateInterval, "semantic state polling interval")
-	fs.DurationVar(&cfg.SemanticEnergyInterval, "semantic-energy-interval", cfg.SemanticEnergyInterval, "semantic energy polling interval")
-	fs.DurationVar(&cfg.SemanticRequestTimeout, "semantic-request-timeout", cfg.SemanticRequestTimeout, "semantic per-request timeout")
-	fs.Func("semantic-read-breaker-failure-budget", "semantic read breaker consecutive failure budget (<=0 disables)", func(value string) error {
-		parsed, err := strconv.Atoi(strings.TrimSpace(value))
-		if err != nil {
-			return fmt.Errorf("invalid semantic-read-breaker-failure-budget %q", value)
-		}
-		cfg.SemanticReadBreakerFailureBudget = parsed
-		cfg.SemanticReadBreakerFailureBudgetSet = true
-		return nil
-	})
-	fs.DurationVar(&cfg.SemanticReadBreakerOpenCooldown, "semantic-read-breaker-open-cooldown", cfg.SemanticReadBreakerOpenCooldown, "semantic read breaker open-state cooldown before probe")
-	fs.IntVar(&cfg.SemanticReadBreakerHalfOpenProbeLimit, "semantic-read-breaker-half-open-probe-limit", cfg.SemanticReadBreakerHalfOpenProbeLimit, "semantic read breaker half-open probes per cooldown window")
-	fs.IntVar(&cfg.SemanticZonePresenceMissThreshold, "semantic-zone-presence-miss-threshold", cfg.SemanticZonePresenceMissThreshold, "consecutive discovery misses required before a zone is marked absent")
-	fs.IntVar(&cfg.SemanticZonePresenceHitThreshold, "semantic-zone-presence-hit-threshold", cfg.SemanticZonePresenceHitThreshold, "consecutive discovery hits required before an absent zone is marked present")
-	fs.DurationVar(&cfg.SemanticDHWStaleTTL, "semantic-dhw-stale-ttl", cfg.SemanticDHWStaleTTL, "maximum age to keep DHW last-known state during cache-sourced/transient failures")
-	fs.DurationVar(&cfg.SemanticRegulatorRecheckInterval, "semantic-regulator-recheck-interval", cfg.SemanticRegulatorRecheckInterval, "regulator capability re-evaluation interval")
-	fs.DurationVar(&cfg.SemanticRegulatorAbsenceGrace, "semantic-regulator-absence-grace", cfg.SemanticRegulatorAbsenceGrace, "grace window before WARN_NO_REGULATOR after regulator disappears")
-	fs.StringVar(&cfg.SemanticCachePath, "semantic-cache-path", cfg.SemanticCachePath, "semantic cache file path for startup preload and live persistence")
-	fs.Func("semantic-interval", "DEPRECATED: semantic state polling interval", func(value string) error {
-		duration, err := time.ParseDuration(value)
-		if err != nil {
-			return fmt.Errorf("invalid semantic-interval %q", value)
-		}
-		cfg.SemanticInterval = duration
-		cfg.SemanticStateInterval = duration
-		return nil
-	})
-	fs.BoolVar(&cfg.BroadcastListen, "broadcast", cfg.BroadcastListen, "enable broadcast listener (separate connection)")
-	fs.StringVar(&cfg.HTTPAddr, "http-addr", cfg.HTTPAddr, "http listen address (empty disables)")
-	fs.StringVar(&cfg.GraphQLPath, "graphql-path", cfg.GraphQLPath, "graphql endpoint path")
-	fs.StringVar(&cfg.SnapshotPath, "snapshot-path", cfg.SnapshotPath, "projection snapshot endpoint path")
-	fs.StringVar(&cfg.SubscriptionPath, "subscription-path", cfg.SubscriptionPath, "graphql subscriptions path")
-	fs.StringVar(&cfg.MCPPath, "mcp-path", cfg.MCPPath, "mcp endpoint path")
-	fs.StringVar(&cfg.UIPath, "ui-path", cfg.UIPath, "portal ui path")
-	fs.StringVar(&cfg.PortalPath, "portal-path", cfg.PortalPath, "dynamic portal path")
-	fs.StringVar(&cfg.DumpUploadPath, "dump-upload-path", cfg.DumpUploadPath, "register dump upload endpoint path")
-	fs.BoolVar(&cfg.MDNSAdvertise, "mdns", cfg.MDNSAdvertise, "advertise graphql endpoint via mdns")
-	fs.StringVar(&cfg.MDNSInstance, "mdns-instance", cfg.MDNSInstance, "mdns instance name")
-	fs.Func("instance-guid", "stable gateway instance UUIDv4 (lowercase)", func(value string) error {
-		normalized, err := normalizeInstanceGUID(value)
-		if err != nil {
-			return err
-		}
-		cfg.InstanceGUID = normalized
-		return nil
-	})
-	fs.StringVar(&cfg.InstanceGUIDSource, "instance-guid-source",
-		cfg.InstanceGUIDSource,
-		"AD27 provenance tag for -instance-guid (runtime_state | legacy_migrated | generated | cli-override); "+
-			"empty defaults to cli-override with a deprecation log when -instance-guid is provided")
-	fs.StringVar(&cfg.RuntimeStatePath, "runtime-state-path",
-		cfg.RuntimeStatePath,
-		"override /data/runtime_state.json path (empty uses runtimestate package default)")
-	fs.StringVar(&cfg.DumpOutputDir, "dump-output-dir", cfg.DumpOutputDir, "unknown device dump output dir")
-	fs.StringVar(&cfg.DumpUploadURL, "dump-upload-url", cfg.DumpUploadURL, "unknown device dump upload url (internal)")
-	fs.BoolVar(&cfg.DumpIncludePII, "dump-include-pii", cfg.DumpIncludePII, "include identifiers in unknown device dumps")
-	fs.BoolVar(&cfg.EnableStaticSeedTable, "enable-static-seed-table", cfg.EnableStaticSeedTable, "plant productids static seed entries (NETX3 0xF1 / 0xF6 / 0x04 / 0xFF, BASV2 0x15 / 0xEC) into registry at startup; default false")
-	fs.BoolVar(&cfg.ObserveFirstEnabled, "observe-first-enabled", cfg.ObserveFirstEnabled, "enable observe-first runtime behavior gates")
-	fs.BoolVar(&cfg.PassiveStateDirectApply, "passive-state-direct-apply", cfg.PassiveStateDirectApply, "allow passive state direct-apply when observe-first is enabled")
-	fs.BoolVar(&cfg.PassiveConfigDirectApply, "passive-config-direct-apply", cfg.PassiveConfigDirectApply, "allow passive config direct-apply when state direct-apply is enabled")
-	fs.Func("external-write-policy", "externally observed write policy: invalidate_only, record_only, or record_and_invalidate", func(value string) error {
-		policy, err := ebusgateway.ParseObserveFirstExternalWritePolicy(value)
-		if err != nil {
-			return err
-		}
-		cfg.ExternalWritePolicy = policy
-		return nil
-	})
-
-	fs.StringVar(&cfg.ProxyListenAddr, "proxy-listen", cfg.ProxyListenAddr, "TCP listen address for ENH proxy clients (e.g. :19001, empty disables)")
-
-	fs.StringVar(&cfg.PhantomInitiatorRejectBytes, "phantom-initiator-reject-bytes",
-		cfg.PhantomInitiatorRejectBytes,
-		"comma-separated hex bytes treated as phantom AND-collision artifacts (e.g. 0x71,0xFD); empty disables filtering. Default 0x71 covers the gateway=0x7F & initiator=0xF1 collision case observed on the live HA bus. Operators on different buses should set this explicitly.")
-
-	fs.Func("source-addr", "source address for scans/semantic reads (e.g. 0xf0, 0x00, or auto)", func(value string) error {
-		value = strings.TrimSpace(strings.ToLower(value))
-		if value == "" {
-			return nil
-		}
-		if value == "auto" {
-			cfg.ScanSource = 0x00
-			cfg.ScanSourceAuto = true
-			return nil
-		}
-		parsed, err := strconv.ParseUint(value, 0, 8)
-		if err != nil {
-			return fmt.Errorf("invalid source-addr %q", value)
-		}
-		cfg.ScanSource = byte(parsed)
-		cfg.ScanSourceAuto = cfg.ScanSource == 0x00
-		return nil
-	})
-	fs.Func("startup-probe-targets", "comma-separated explicit startup directed-probe targets (e.g. 0x08,0x15)", func(value string) error {
-		targets, err := parseStartupProbeTargets(value)
-		if err != nil {
-			return err
-		}
-		cfg.StartupProbeTargets = targets
-		return nil
-	})
-	fs.Func("startup-source-override", "override source address for source-selection-capable direct transports (e.g. 0xf0)", func(value string) error {
-		value = strings.TrimSpace(strings.ToLower(value))
-		if value == "" {
-			cfg.StartupSource.Source = nil
-			return nil
-		}
-		parsed, err := strconv.ParseUint(value, 0, 8)
-		if err != nil {
-			return fmt.Errorf("invalid startup-source-override %q", value)
-		}
-		source := uint8(parsed)
-		if source == 0x00 {
-			return fmt.Errorf("invalid startup-source-override %q: source 0x00 is not a valid active initiator", value)
-		}
-		cfg.StartupSource.Source = &source
-		return nil
-	})
-	fs.BoolVar(&cfg.StartupSource.Validate, "startup-source-override-validate", cfg.StartupSource.Validate, "run source-address selector in advisory-only mode alongside startup-source-override")
-	return inputs
-}
-
-// parseHexByteList parses a comma-separated list of hex byte literals
-// (e.g. "0x71, 0xFD" or "71,FD") into []byte. Empty input returns a nil
-// slice with no error. Whitespace around items is trimmed. Per-byte
-// parsing accepts the standard Go base prefix (0x / 0X / no-prefix-hex),
-// so "0x71", "0X71", and "71" all parse to 0x71. Invalid items return an
-// error naming the offending substring.
-//
-// Used to materialize the --phantom-initiator-reject-bytes CLI flag into
-// the IsKnownInitiatorByte predicate plumbed into adaptermux (Codex P2
-// thread on PR #634, batch-24).
-func parseHexByteList(value string) ([]byte, error) {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return nil, nil
-	}
-	parts := strings.Split(value, ",")
-	out := make([]byte, 0, len(parts))
-	for _, part := range parts {
-		part = strings.TrimSpace(part)
-		if part == "" {
-			continue
-		}
-		// strconv.ParseUint with base 0 accepts "0x71", "0X71", "71"
-		// (treated as decimal!) — to keep "71" parsing as hex we
-		// fall back to base 16 when no 0x prefix is present.
-		var (
-			parsed uint64
-			err    error
-		)
-		lower := strings.ToLower(part)
-		if strings.HasPrefix(lower, "0x") {
-			parsed, err = strconv.ParseUint(part[2:], 16, 8)
-		} else {
-			parsed, err = strconv.ParseUint(part, 16, 8)
-		}
-		if err != nil {
-			return nil, fmt.Errorf("invalid hex byte %q: %w", part, err)
-		}
-		out = append(out, byte(parsed))
-	}
-	return out, nil
-}
-
-func parseStartupProbeTargets(value string) ([]byte, error) {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return nil, nil
-	}
-	parts := strings.FieldsFunc(value, func(r rune) bool {
-		return r == ',' || r == ' ' || r == '\t' || r == '\n'
-	})
-	targets := make([]byte, 0, len(parts))
-	seen := make(map[byte]struct{}, len(parts))
-	for _, part := range parts {
-		part = strings.TrimSpace(strings.ToLower(part))
-		if part == "" {
-			continue
-		}
-		parsed, err := strconv.ParseUint(part, 0, 8)
-		if err != nil {
-			return nil, fmt.Errorf("invalid startup-probe-targets address %q", part)
-		}
-		target := byte(parsed)
-		if target < 0x03 || target >= 0xFE || target == 0xAA || target == 0xA9 || isInitiatorCapableAddress(target) {
-			return nil, fmt.Errorf("startup-probe-targets address 0x%02X outside target-capable range", target)
-		}
-		if _, ok := seen[target]; ok {
-			continue
-		}
-		seen[target] = struct{}{}
-		targets = append(targets, target)
-	}
-	return targets, nil
-}
-
-// wireAdapterDirect creates and starts the adapter multiplexer if the
-// transport protocol is adapter-direct. It configures both active and
-// passive transports in cfg before gateway construction.
-//
-// Returns a closer function for the multiplexer, the instance-scoped
-// activeTxnClassifier (the mux), or nils if not in adapter-direct mode.
-// The classifier is threaded explicitly into startDiscoveryScanLoopFn
-// at the call site in run() — never captured in a package-level closure —
-// so classifier state is strictly instance-local (Codex PR #502 P2).
-func wireAdapterDirect(ctx context.Context, cfg *ebusgateway.Config) (func() error, activeTxnClassifier, error) {
-	return wireAdapterDirectWithConnectionLost(ctx, cfg, nil)
-}
-
-const adapterDirectENSProtocol ebusgateway.TransportProtocol = "adapter-direct-ens"
-
-func wireAdapterDirectWithConnectionLost(ctx context.Context, cfg *ebusgateway.Config, onConnectionLost func()) (func() error, activeTxnClassifier, error) {
-	normalized, err := ebusgateway.NormalizeEBusDriverTransportConfig(cfg.TransportConfig)
-	if err != nil {
-		return nil, nil, err
-	}
-	if normalized.Protocol != ebusgateway.TransportAdapterDirect && normalized.Protocol != adapterDirectENSProtocol {
-		return nil, nil, nil
-	}
-	cfg.TransportConfig = normalized
-	network := normalized.Network
-	address := normalized.Address
-	if address == "" {
-		return nil, nil, fmt.Errorf("adapter-direct requires an address (e.g. adapter-direct://boiler.local:9999): %w", ebuserrors.ErrInvalidPayload)
-	}
-
-	// Determine ENH vs ENS sub-protocol. ENH is the default.
-	// The shared normalizer preserves adapter-direct-ens as a private
-	// canonical literal after stripping the URI.
-	adapterProtocol := adapterDirectMuxProtocol(normalized.Protocol)
-
-	muxCfg := adaptermux.Config{
-		Protocol:     adapterProtocol,
-		Network:      network,
-		Address:      address,
-		DialTimeout:  cfg.TransportConfig.DialTimeout,
-		ReadTimeout:  cfg.TransportConfig.ReadTimeout,
-		WriteTimeout: cfg.TransportConfig.WriteTimeout,
-		// F-30 (batch-27, iter7, 2026-05-14; batch-24 round-5
-		// parameterization, 2026-05-17): wire IsKnownInitiatorByte to
-		// filter bit-arbitration phantom bytes from external session
-		// forwarding. When the gateway loses arbitration to another
-		// initiator, the adapter reports FAILED with the bit-wise-AND
-		// result of the colliding initiators. This AND result is often
-		// NOT a real initiator on the bus (e.g., 0x7F & 0xF1 = 0x71,
-		// where no 0x71 initiator exists on the observed bus).
-		//
-		// Pre-F-30: phantom bytes were forwarded to ebusd as ENH_RES_FAILED
-		// data; ebusd's bus reconstructor interpreted each phantom as a
-		// real frame source and advanced its state machine to expect a
-		// frame from that fictitious initiator. The next real wire bytes
-		// (from the actual winning initiator) then mismatched, leaving
-		// ebusd's state in bs_recvCmd/bs_skip — the "arbitration won in
-		// invalid state" trigger when ebusd's NEXT grant arrives.
-		//
-		// Iter6 forensic measured 58 invalid-state events / 5 min
-		// despite F-28+F-29 cooldowns reducing the cause-1 (timeout)
-		// path. Cause-2 (state mismatch from phantom forwarding) is now
-		// the dominant residual.
-		//
-		// Iter7 (hardcoded): rejected the SPECIFIC phantom 0x71 observed
-		// on this bus. Round-5 (batch-24) generalizes via the
-		// --phantom-initiator-reject-bytes CLI flag — default "0x71"
-		// preserves iter7 behavior; operators on different buses should
-		// set this explicitly (empty disables filtering, CSV extends
-		// rejection to additional phantoms). The runtime_state-backed
-		// lookup originally planned for iter8 is still on the table but
-		// requires bus-membership convergence first.
-		//
-		// Returning false on the predicate routes the FAILED byte
-		// through the gateway's suppression path (mux.go ~1523): the
-		// byte is NOT delivered to external sessions, the per-session
-		// notify gets the bidder's own initiator instead of the
-		// phantom, and the passive emit / logging still fires for
-		// observability.
-	}
-	phantomBytes, err := parseHexByteList(cfg.PhantomInitiatorRejectBytes)
-	if err != nil {
-		return nil, nil, fmt.Errorf("invalid -phantom-initiator-reject-bytes %q: %w", cfg.PhantomInitiatorRejectBytes, err)
-	}
-	if len(phantomBytes) == 0 {
-		// Empty CSV disables filtering entirely — every byte is
-		// "known", so the suppression path in mux.go never fires.
-		muxCfg.IsKnownInitiatorByte = nil
-	} else {
-		// Copy the slice into the closure so subsequent mutations of
-		// the caller's slice (none today, but defensive) can't shift
-		// the predicate's behavior at runtime.
-		pb := append([]byte(nil), phantomBytes...)
-		muxCfg.IsKnownInitiatorByte = func(b byte) bool {
-			for _, r := range pb {
-				if r == b {
-					return false
-				}
-			}
-			return true
-		}
-	}
-	if muxCfg.DialTimeout == 0 {
-		muxCfg.DialTimeout = 5 * time.Second
-	}
-	// Mux read timeout controls how often the idle-timeout branch runs
-	// (tryGrantAndStart) AND how quickly activeCh receives bytes from
-	// the upstream ENH transport. 50ms matches the eBUS SYN interval so
-	// waitForSyn (which needs 2 SYNs) completes within ~100ms instead
-	// of ~400ms at 200ms, preventing scan-abort after collision retry.
-	muxCfg.ReadTimeout = 50 * time.Millisecond
-	if muxCfg.WriteTimeout == 0 {
-		muxCfg.WriteTimeout = 5 * time.Second
-	}
-
-	// Phase 3 Step B3.7 (Codex round-1 MAJOR on PR #650): wire the
-	// HELIANTHUS_V8_CLASSIFIER_MODE env var into the mux config.
-	// Without this read, the v8 classifier is unreachable in
-	// production — Config.V8ClassifierMode would always be the
-	// zero-value (ModeOff) regardless of what operators set in
-	// the addon options. ParseMode handles case-insensitive
-	// "off|shadow|enforce" plus convenience synonyms ("disabled",
-	// "false", etc.); an unset env var defaults to ModeOff (safe);
-	// an unrecognized value fails loudly via the returned error.
-	if envMode := os.Getenv(v8classifier.EnvVarName); envMode != "" {
-		parsedMode, err := v8classifier.ParseMode(envMode)
-		if err != nil {
-			return nil, nil, fmt.Errorf("invalid %s=%q (expected off|shadow|enforce): %w", v8classifier.EnvVarName, envMode, err)
-		}
-		muxCfg.V8ClassifierMode = parsedMode
-		if parsedMode != v8classifier.ModeOff {
-			log.Printf("v8 classifier enabled: mode=%s", parsedMode)
-		}
-	}
-
-	mux := adaptermux.New(muxCfg)
-	connectionLostOwner := onConnectionLost
-	if onConnectionLost != nil {
-		var connectionLostOnce sync.Once
-		connectionLostOwner = func() { connectionLostOnce.Do(onConnectionLost) }
-	}
-	mux.SetConnectionLostCallback(connectionLostOwner)
-	// Codex PR #502 P2: the instance-scoped classifier is returned to
-	// run() and threaded explicitly as the 5th argument to
-	// startDiscoveryScanLoopFn. No package-level closure captures this
-	// mux — so a second wireAdapterDirect call (e.g. in the same process
-	// across tests) sees its own classifier at the call site.
-
-	// Create passive transport BEFORE Start() so the callback is wired.
-	// Only create it when BroadcastListen is enabled — otherwise no
-	// consumer reads from the passive channel and reset delivery blocks
-	// readLoop (P0 deadlock fix).
-	if cfg.BroadcastListen {
-		cfg.PassiveTransport = mux.PassiveTransport()
-
-		// Skip passive warmup in adapter-direct mode. The passive stream
-		// only sees third-party traffic (gateway bytes are suppressed by
-		// the mux), so warmup thresholds are never met organically. Zero
-		// thresholds tell the store to promote immediately.
-		cfg.ObserveFirstWarmupCompletedTransactions = 0
-		cfg.ObserveFirstWarmupConnectedWindow = 0
-		cfg.ObserveFirstWarmupPostResetTransactions = 0
-		cfg.ObserveFirstWarmupPostResetWindow = 0
-	}
-
-	if err := mux.Start(ctx); err != nil {
-		return nil, nil, fmt.Errorf("start multiplexer: %w", err)
-	}
-
-	log.Printf("adapter-direct: connected to %s/%s", network, address)
-
-	// Start proxy listener if configured (exposes ENH endpoint for
-	// external clients like ebusd).
-	var proxyListener *adaptermux.ProxyListener
-	if cfg.ProxyListenAddr != "" {
-		onProxyFatal := adapterDirectProxyFatalCallback(mux, connectionLostOwner)
-		pl, err := adaptermux.NewProxyListenerWithFatalCallback(ctx, mux, cfg.ProxyListenAddr, log.Default(), onProxyFatal)
-		if err != nil {
-			if cerr := mux.Close(); cerr != nil {
-				log.Printf("adapter-direct: mux.Close after proxy-listener error: %v", cerr)
-			}
-			return nil, nil, fmt.Errorf("proxy listener: %w", err)
-		}
-		proxyListener = pl
-		log.Printf("adapter-direct: proxy listener on %s", pl.Addr())
-	}
-
-	// Configure gateway transports.
-	cfg.Transport = mux.ActiveTransport()
-
-	// Return a closer that cleans up both the proxy listener (if any)
-	// and the mux itself. This covers early run() failures where
-	// gateway.Close() never runs (and thus activeTransport.Close
-	// never calls mux.Close). On normal shutdown the gateway's
-	// transport.Close calls mux.Close first — that is safe because
-	// mux.Close is idempotent (sync.Once guarded).
-	closer := func() error {
-		if proxyListener != nil {
-			if err := proxyListener.Close(); err != nil {
-				log.Printf("adapter-direct: proxyListener.Close: %v", err)
-			}
-		}
-		return mux.Close()
-	}
-	return closer, activeTxnClassifier(mux), nil
-}
-
-func adapterDirectProxyFatalCallback(mux *adaptermux.Mux, owner func()) func(error) {
-	if mux == nil || owner == nil {
-		return nil
-	}
-	return func(err error) {
-		// Fatal listener loss is a generation boundary just like upstream loss.
-		// Fence and synchronously retire every existing proxy session before
-		// publishing failure; wireAdapterDirect's shared one-shot owner prevents
-		// duplicate recovery with a concurrent upstream loss.
-		mux.RetireManagedProxySessions()
-		log.Printf("adapter-direct: proxy listener failed: %v", err)
-		owner()
-	}
-}
-
-func adapterDirectMuxProtocol(protocol ebusgateway.TransportProtocol) string {
-	if protocol == adapterDirectENSProtocol {
-		return "ens"
-	}
-	return "enh"
-}
-
-func adapterDirectProxyEnabled(config ebusgateway.TransportConfig) bool {
-	return ebusgateway.EBusDriverTransportProtocol(config) == ebusgateway.TransportAdapterDirect
-}
-
-// v8AdminEventsResponse is the JSON envelope returned by
-// /debug/v8/admin-events. Stable wire format — operator tooling
-// (dashboards, classifier-tuning workflows) parse this.
-//
-// Fields:
-//
-//   - Events: copy of all events the ring buffer held at drain
-//     time, in FIFO order (oldest first). Each event includes the
-//     wire byte, FSM state at decision time, escape-decoded
-//     provenance, and monotonic-clock timestamp.
-//
-//   - Dropped: number of events the ring buffer dropped (oldest-
-//     first FIFO) since the last drain. Non-zero indicates the
-//     consumer is polling too slowly for the production event
-//     rate; saturate response is acceptable but the operator
-//     should tighten the poll cadence.
-//
-//   - Kind / FSMState: JSON-encoded as their stable string labels
-//     via the v8AdminEventJSON intermediate so the wire format
-//     doesn't expose numeric enum values that could shift across
-//     gateway versions.
-type v8AdminEventsResponse struct {
-	Events  []v8AdminEventJSON `json:"events"`
-	Dropped uint64             `json:"dropped"`
-}
-
-// v8AdminEventJSON is the per-event wire shape returned by
-// /debug/v8/admin-events. Mirrors v8classifier.ClassifierAdminEvent
-// but encodes the enums as stable string labels.
-type v8AdminEventJSON struct {
-	// At is the monotonic-clock observation time of the event.
-	// Encoded as RFC3339Nano for human-readable parsing without
-	// loss of resolution.
-	At time.Time `json:"at"`
-
-	// Kind is the v8classifier.AdminEventKind label (e.g.
-	// "protocol_fault", "aa_injection_drop").
-	Kind string `json:"kind"`
-
-	// FSMState is the telegram_fsm.State label at the time of
-	// the event (e.g. "MASTER_HEADER", "MASTER_PAYLOAD").
-	FSMState string `json:"fsm_state"`
-
-	// Byte is the wire byte that triggered the event, encoded
-	// as a 0x-prefixed two-digit hex string for human
-	// readability (e.g. "0xAA"). Zero-valued for non-byte
-	// events (queue overflow, echo timeouts).
-	Byte string `json:"byte"`
-
-	// WasEscaped reports whether the byte arrived as an
-	// escape-decoded payload (true) or as a raw wire byte
-	// (false). The shadow→enforce promotion gate uses this to
-	// distinguish true-positive AA-injection (raw wire 0xAA
-	// mid-frame) from data-byte 0xAA that arrived legitimately
-	// via the ENS 0xA9 escape sequence.
-	WasEscaped bool `json:"was_escaped"`
-}
-
-// handleV8AdminEvents returns the active v8 classifier's admin
-// event ring buffer as JSON.
-//
-// **Default GET drains the ring** (destructive). Each successful
-// GET returns the events accumulated since the last drain and
-// EMPTIES the buffer. This is the documented contract for
-// long-running poller tooling. Operators MUST poll at a steady
-// cadence (every few seconds in production); a stuck consumer
-// drops the OLDEST events FIFO and the `dropped` field surfaces
-// the saturation.
-//
-// **`?peek=true` returns events WITHOUT draining** (non-
-// destructive). For ad-hoc operator inspection (`curl | jq`,
-// browser visit, dashboards, health checks) that should NOT
-// consume the long-running poller's evidence stream. The full
-// ring contents are returned every time — the operator is
-// responsible for de-duplicating against prior peeks. Codex
-// round-1 LOW on PR #657 — the GET-drain footgun when
-// concurrent consumers share the HTTP surface.
-//
-// Nil-safe: when no classifier is registered (non-adapter-direct
-// transport, or run() not yet completed), returns
-// `{"events": [], "dropped": 0}` so probing tooling sees a
-// well-formed response on every transport rather than 404-ing.
-//
-// Only GET is accepted — POST/etc respond 405 to avoid surprising
-// side effects.
-func handleV8AdminEvents(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodGet {
-		w.Header().Set("Allow", "GET")
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-
-	peekOnly := r.URL.Query().Get("peek") == "true"
-
-	resp := v8AdminEventsResponse{
-		Events: []v8AdminEventJSON{},
-	}
-
-	withCurrentV8AdminClassifier(func(classifier *v8classifier.Classifier) {
-		var events []v8classifier.ClassifierAdminEvent
-		var dropped uint64
-		if peekOnly {
-			// Non-destructive ATOMIC read: events + dropped come
-			// from a single mutex acquire (Codex round-2 MEDIUM
-			// on PR #657 — separate calls would race against a
-			// concurrent drain/overflow emit). The classifier
-			// exposes the destructive variant via DrainAdminEvents.
-			events, dropped = classifier.PeekAdminEvents()
-		} else {
-			events, dropped = classifier.DrainAdminEvents()
-		}
-		resp.Dropped = dropped
-		if len(events) > 0 {
-			resp.Events = make([]v8AdminEventJSON, len(events))
-			for i, ev := range events {
-				resp.Events[i] = v8AdminEventJSON{
-					At:         ev.At,
-					Kind:       ev.Kind.String(),
-					FSMState:   ev.FSMState.String(),
-					Byte:       fmt.Sprintf("0x%02X", ev.Byte),
-					WasEscaped: ev.WasEscaped,
-				}
-			}
-		}
-	})
-
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	if peekOnly {
-		// Peek is idempotent. `private` keeps shared
-		// intermediary caches (CDNs, reverse proxies) from
-		// storing the response on behalf of unrelated
-		// operators — this is a debug surface, not public
-		// content. `max-age=1` keeps tooling that polls every
-		// few seconds from seeing stale data while still
-		// allowing the browser's own cache to debounce
-		// duplicate-refresh storms. Codex round-2 LOW on
-		// PR #657.
-		w.Header().Set("Cache-Control", "private, max-age=1")
-	} else {
-		// Drain is destructive, the same response is never
-		// valid twice. Tools that cache the response would
-		// silently mask new events.
-		w.Header().Set("Cache-Control", "no-store")
-	}
-	if err := json.NewEncoder(w).Encode(resp); err != nil {
-		log.Printf("/debug/v8/admin-events: encode error: %v", err)
-	}
-}
-
-func startHTTPServer(
-	ctx context.Context,
-	cfg ebusgateway.Config,
-	gateway *ebusgateway.Gateway,
-	builder *graphql.Builder,
-	hub *graphql.BroadcastHub,
-	semanticProvider graphql.SemanticProvider,
-	eebusProvider mcp.EEBusV1Provider,
-	eebusCommandRouter mcp.EEBusV1CommandRouter,
-	modbusProvider mcp.ModbusV1Provider,
-	scheduleWriter mcp.ScheduleWriter,
-	configWriter mcp.ConfigWriter,
-	busObservability *ebusgateway.BusObservabilityStore,
-	watchSummaryProvider mcp.WatchSummaryProvider,
-	eebusAdminHandler http.Handler,
-	eebusLifecycle *eebusRuntimeLifecycle,
-	ebusProxyReadiness func() string,
-	ebusSourceProvider func() (byte, bool),
-	buildInfo gatewayBuildInfo,
-	ebusDriver *ebusDriverController,
-) (*http.Server, mdns.Advertiser, error) {
-	if cfg.HTTPAddr == "" {
-		return nil, nil, nil
-	}
-	if gateway == nil {
-		return nil, nil, fmt.Errorf("gateway missing for http server")
-	}
-	if builder == nil {
-		return nil, nil, fmt.Errorf("graphql builder missing for http server")
-	}
-	if hub == nil {
-		return nil, nil, fmt.Errorf("graphql broadcast hub missing for http server")
-	}
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if err := cfg.ValidatePortalPV(); err != nil {
-		return nil, nil, fmt.Errorf("validate Portal PV configuration: %w", err)
-	}
-	portalPVClient, err := newPortalPVClient(cfg.PortalPV)
-	if err != nil {
-		return nil, nil, fmt.Errorf("portal PV configuration: %w", err)
-	}
-
-	queryHandler, err := graphql.NewInvokeHandler(builder, gateway.Registry, gateway.Router)
-	if err != nil {
-		return nil, nil, err
-	}
-	snapshotHandler, err := graphql.NewProjectionSnapshotHandler(builder)
-	if err != nil {
-		return nil, nil, err
-	}
-	subscriptionHandler, err := graphql.NewSubscriptionHandler(builder, gateway.Registry, gateway.Router, hub)
-	if err != nil {
-		return nil, nil, err
-	}
-	mcpServer, err := mcp.NewServer(gateway.Registry, gateway.Router)
-	if err != nil {
-		return nil, nil, err
-	}
-	if err := mcpServer.SetServerVersion(buildInfo.ReleaseVersion); err != nil {
-		return nil, nil, fmt.Errorf("configure MCP build identity: %w", err)
-	}
-	if eebusProvider != nil {
-		if err := mcpServer.RegisterEEBusV1Provider(eebusProvider); err != nil {
-			return nil, nil, fmt.Errorf("register eeBUS MCP provider: %w", err)
-		}
-	}
-	if eebusCommandRouter != nil {
-		if err := mcpServer.RegisterEEBusV1CommandRouter(eebusCommandRouter); err != nil {
-			return nil, nil, fmt.Errorf("register eeBUS MCP command router: %w", err)
-		}
-	}
-	mcp.RegisterModbusV1Tools(mcpServer, modbusProvider)
-	if source, ok := eebusProvider.(mcp.LeafPromotionCaptureSource); ok {
-		if capture := source.LeafPromotionCapture(); capture != nil {
-			if err := mcpServer.RegisterLeafPromotionCapture(capture); err != nil {
-				return nil, nil, fmt.Errorf("register leaf promotion capture: %w", err)
-			}
-		}
-	}
-	oneShotRuntime, err := newSynchronizedEvidenceOneShotRuntime(
-		cfg.EvidenceOneShotEnabled,
-		eebusProvider,
-		eebusCommandRouter,
-		buildInfo,
-	)
-	if err != nil {
-		return nil, nil, err
-	}
-	if oneShotRuntime != nil {
-		if err := mcpServer.RegisterSynchronizedEvidenceCapture(oneShotRuntime); err != nil {
-			return nil, nil, fmt.Errorf("register synchronized evidence capture: %w", err)
-		}
-	}
-	mcpServer.SetAdmittedRPCSourceProvider(ebusSourceProvider)
-	mcpServer.SetStatusProvider(newMCPRuntimeStatusProvider(semanticProvider, ebusSourceProvider))
-	if busObservability != nil {
-		mcpServer.SetBusObservabilityProvider(newMCPBusObservabilityProvider(busObservability))
-	}
-	if watchSummaryProvider != nil {
-		mcpServer.SetWatchSummaryProvider(watchSummaryProvider)
-	}
-	mcpServer.SetSemanticProvider(newMCPSemanticProvider(semanticProvider))
-
-	// M2a_GATEWAY_MCP (execution-plans#19): install Vaillant B503 MCP tool
-	// surface. Uses a deferred dispatcher stub — production B524-style raw
-	// RPC wiring for the 2-byte (family, selector) frame is scheduled as a
-	// follow-up under the M2b / M3 rollout (the MCP substrate currently
-	// models dispatch as (plane, method, params) through the catalog, and
-	// B503 has not yet been added to the catalog — intentional, per plan
-	// AD01: Vaillant protocol knowledge stays out of ebusreg).
-	//
-	// Registering here ensures:
-	//   - `ebus.v1.vaillant.*` tools are listed by tools/list (P1 lint from
-	//     Codex review of M2a — tools MUST be reachable by clients, even
-	//     if production dispatch is not yet wired);
-	//   - capability signal is emitted;
-	//   - forbidden-surface guards (TestNoVaillantInstallWriteTools) apply
-	//     to the production build, not just the test harness.
-	//
-	// With the stub dispatcher, read tools surface `UPSTREAM_RPC_FAILED`
-	// with the "production wiring pending" message; live-monitor action
-	// paths use the real session FSM so EXPIRED normalization, session
-	// epochs, and owner-conditional release are all exercised — only the
-	// raw bus dispatch is stubbed.
-	b503rt := installVaillantB503(mcpServer, gateway, &cfg, ebusSourceProvider)
-	if b503rt != nil && ebusDriver != nil {
-		ebusDriver.SetLifecycleObserver(b503rt)
-	}
-	// M2b_GATEWAY_GRAPHQL (execution-plans#19): wire the GraphQL B503
-	// provider to the same Manager + Dispatcher the MCP surface uses. A
-	// single Manager across both surfaces is mandatory — GraphQL
-	// Enable/Read/Disable operating on a separate Manager would break
-	// the single-owner session invariant.
-	if b503rt != nil {
-		builder.SetVaillantB503Provider(newB503GraphQLProvider(b503rt))
-	}
-
-	// M4c2: populate the package-level responder capability provider
-	// based on the active transport protocol. Consumers apply fail-closed
-	// semantics on absence, so this MUST be called before any MCP
-	// surface serves its first envelope. The provider closure is
-	// evaluated per-envelope; hot-path cost is a single pointer load +
-	// struct copy. See decision doc @ 567a6798 §4.2 + §5.
-	//
-	// A nil return from buildResponderCapabilityProvider means the raw
-	// transport protocol does not canonicalize to any of the three
-	// locked rows at v1.1 (ENH / ENS / ebusd-tcp). In that case we omit
-	// the capability entirely so consumers fall back to §4.3 rule 1
-	// (absence → scope=none, fail-closed). This preserves invariant I1
-	// (exactly three rows at v1.1) and I2 (active.transport MUST appear
-	// in transports[]) without widening the schema.
-	// Pass the live transport instance (gateway.Transport) so the provider
-	// can type-assert against ebusgoTransport.ResponderTransport. The
-	// adapter-direct mux returns a RawTransport that does NOT satisfy
-	// ResponderTransport, so config-string "enh" + mux active path
-	// correctly downgrades to scope=none (see Codex P1 on PR #509).
-	if provider := buildResponderCapabilityProvider(cfg, gateway.Transport); provider != nil {
-		ebus_standard.SetResponderCapabilityProvider(provider)
-	} else {
-		log.Printf("warning: meta.capabilities.responder omitted: transport protocol %q does not canonicalize to ENH/ENS/ebusd-tcp", cfg.TransportConfig.Protocol)
-	}
-	if scheduleWriter != nil {
-		mcpServer.SetScheduleWriter(scheduleWriter)
-	}
-	if configWriter != nil {
-		mcpServer.SetConfigWriter(configWriter)
-	}
-
-	mux := http.NewServeMux()
-	if busObservability != nil {
-		mux.Handle(normalizeMountPath(cfg.MetricsPath, ebusgateway.DefaultMetricsPath), busObservability.MetricsHandler())
-	}
-	// Expose expvar surfaces (including the 11 startup_source_selection_* counters
-	// from M5) via /debug/vars. The expvar package's init registers the
-	// handler on http.DefaultServeMux, but the gateway uses its own mux so
-	// the handler must be wired explicitly. Resolves cruise-run #20
-	// validation finding: M5 expvars were defined + Publish()'d but not
-	// reachable over HTTP.
-	mux.Handle("/debug/vars", expvar.Handler())
-	// F-NEW-26 (2026-05-21): expose the v8 classifier's admin
-	// event ring buffer over HTTP so operators can introspect the
-	// byte-level detail behind the helianthus_v8_shadow_would_have_dropped_total
-	// counter. Without this endpoint the aggregate counter is a
-	// black box: an operator running v8 in shadow mode sees the
-	// count rising but cannot distinguish true-positive (real wire
-	// AA-injection that SHOULD be filtered when promoted to enforce)
-	// from false-positive (legitimate traffic v8 over-eagerly flags).
-	// The per-event detail (byte value + FSM state + escape
-	// provenance + timestamp) is the operator's evidence base for
-	// the shadow→enforce promotion gate documented in
-	// helianthus-docs-ebus/deployment/prometheus-alerts.md.
-	//
-	// Endpoint contract:
-	//   GET /debug/v8/admin-events
-	//   Response: application/json
-	//   Body: {"events": [<ClassifierAdminEvent>...], "dropped": <uint64>}
-	//
-	// DRAIN semantics: each successful GET returns the events
-	// accumulated since the last drain and EMPTIES the ring
-	// buffer. Operators MUST poll at a steady cadence (every
-	// few seconds in production); a stuck consumer drops the
-	// OLDEST events FIFO and the "dropped" field surfaces the
-	// saturation.
-	//
-	// Nil-safe: when the classifier is unset (non-adapter-direct
-	// transport, or run() not yet completed), the handler returns
-	// `{"events": [], "dropped": 0}` — the surface stays
-	// available so tooling that probes it unconditionally does
-	// not 404.
-	mux.HandleFunc("/debug/v8/admin-events", handleV8AdminEvents)
-	mux.Handle(cfg.GraphQLPath, queryHandler)
-	mux.Handle(cfg.SnapshotPath, snapshotHandler)
-	mux.Handle(cfg.SubscriptionPath, subscriptionHandler)
-	mux.Handle(cfg.MCPPath, mcpServer.Handler())
-	mux.Handle("/admin/eebus/v1/", eebusAdminHandler)
-	if cfg.DumpUploadPath != "" {
-		uploadPath := cfg.DumpUploadPath
-		if !strings.HasPrefix(uploadPath, "/") {
-			uploadPath = "/" + uploadPath
-		}
-		mux.Handle(uploadPath, ebusgateway.NewRegisterDumpUploadHandler(cfg.DumpOutputDir))
-	}
-	if cfg.UIPath != "" {
-		uiPath := normalizeMountPath(cfg.UIPath, "/ui")
-		uiHandler := ui.NewHandler(cfg.GraphQLPath)
-		mux.Handle(uiPath+"/", http.StripPrefix(uiPath, uiHandler))
-		mux.HandleFunc(uiPath, func(w http.ResponseWriter, r *http.Request) {
-			http.Redirect(w, r, uiPath+"/", http.StatusMovedPermanently)
-		})
-	}
-	if cfg.PortalPath != "" {
-		portalPath := normalizeMountPath(cfg.PortalPath, "/portal")
-		var getPortalBusObservability func() any
-		if busObservability != nil {
-			getPortalBusObservability = func() any {
-				return busObservability.Snapshot().Summary
-			}
-		}
-		portalHandler := portal.NewHandler(portal.Options{
-			GraphQLPath:      cfg.GraphQLPath,
-			SnapshotPath:     cfg.SnapshotPath,
-			SubscriptionPath: cfg.SubscriptionPath,
-			MCPPath:          cfg.MCPPath,
-			EEBusAdminPath:   "/admin/eebus/v1",
-			Readiness: func() portal.RuntimeReadiness {
-				return projectGatewayReadiness(ebusProxyReadiness, eebusLifecycle.LifecycleSnapshot())
-			},
-			GatewayVersion:    buildInfo.ReleaseVersion,
-			BuildID:           buildInfo.BuildID,
-			SemanticPVEnabled: cfg.PortalPV.SemanticEnabled,
-			RawModbusEnabled:  cfg.PortalPV.RawReadEnabled,
-			SemanticPV:        portalPVClient,
-			ModbusProvider:    modbusProvider,
-			RawModbusAudit: func(event portal.RawModbusAuditEvent) {
-				log.Printf("portal_modbus_raw_audit request_id=%s surface=%s tool=%s unit_id=%s function=%s offset=%s quantity=%s outcome=%s error_code=%s duration_ms=%d endpoint_ref=%s timestamp=%s",
-					event.RequestID, event.Surface, event.Tool, auditNumber(event.UnitID), auditNumber(event.Function),
-					auditNumber(event.Offset), auditNumber(event.Quantity), event.Outcome, event.ErrorCode,
-					event.DurationMS, event.EndpointRef, event.At.Format(time.RFC3339Nano))
-			},
-			ListRegistry: func() []portal.RegistryDevice {
-				schemaSnapshot := builder.FreshSchema()
-				schemaByAddr := make(map[byte]graphql.Device, len(schemaSnapshot.Devices))
-				for _, sd := range schemaSnapshot.Devices {
-					schemaByAddr[sd.Address] = sd
-				}
-				items := make([]portal.RegistryDevice, 0)
-				gateway.Registry.Iterate(func(entry registry.DeviceEntry) bool {
-					if entry == nil {
-						return true
-					}
-					rawAddrs := entry.Addresses()
-					intAddrs := make([]int, len(rawAddrs))
-					for i, a := range rawAddrs {
-						intAddrs[i] = int(a)
-					}
-					device := portal.RegistryDevice{
-						Address:      int(entry.PrimaryDisplayAddress()),
-						Addresses:    intAddrs,
-						Manufacturer: entry.Manufacturer(),
-						DeviceID:     entry.DeviceID(),
-						SerialNumber: entry.SerialNumber(),
-						Software:     entry.SoftwareVersion(),
-						Hardware:     entry.HardwareVersion(),
-						Planes:       make([]portal.RegistryPlane, 0),
-					}
-					if sd, ok := schemaByAddr[entry.PrimaryDisplayAddress()]; ok {
-						device.DisplayName = sd.DisplayName
-						device.Role = sd.Role
-					}
-					for _, plane := range entry.Planes() {
-						if plane == nil {
-							continue
-						}
-						methods := plane.Methods()
-						methodNames := make([]string, 0, len(methods))
-						for _, method := range methods {
-							if method == nil {
-								continue
-							}
-							methodNames = append(methodNames, method.Name())
-						}
-						device.Planes = append(device.Planes, portal.RegistryPlane{
-							Name:    plane.Name(),
-							Methods: methodNames,
-						})
-					}
-					items = append(items, device)
-					return true
-				})
-				return items
-			},
-			ListSemantic: func() portal.SemanticSnapshot {
-				if semanticProvider == nil {
-					return portal.SemanticSnapshot{}
-				}
-				verdict := portalFM5Interpretation(semanticProvider)
-				var degradedReason *string
-				if verdict.DegradedReason != "" {
-					reason := string(verdict.DegradedReason)
-					degradedReason = &reason
-				}
-				return portal.SemanticSnapshot{
-					Zones:               mapPortalZones(semanticProvider.Zones()),
-					DHW:                 mapPortalDHW(semanticProvider.DHW()),
-					Energy:              mapPortalEnergyTotals(semanticProvider.EnergyTotals()),
-					BoilerStatus:        mapPortalBoilerStatus(semanticProvider.BoilerStatus()),
-					System:              mapPortalSystemStatus(semanticProvider.System()),
-					Circuits:            mapPortalCircuits(semanticProvider.Circuits()),
-					RadioDevices:        mapPortalRadioDevices(semanticProvider.RadioDevices()),
-					FM5Mode:             string(verdict.Mode),
-					FM5DegradedReason:   degradedReason,
-					FM5EvidenceRevision: verdict.EvidenceRevision,
-					Solar:               mapPortalSolarStatus(semanticProvider.Solar()),
-					Cylinders:           mapPortalCylinders(semanticProvider.Cylinders()),
-					AdapterInfo:         mapPortalAdapterInfo(semanticProvider.AdapterHardwareInfo()),
-					CapturedUTC:         time.Now().UTC().Format(time.RFC3339),
-				}
-			},
-			GetBusObservability: getPortalBusObservability,
-			ListProjections: func() []portal.ProjectionDevice {
-				snapshot := builder.FreshSchema()
-				items := make([]portal.ProjectionDevice, 0, len(snapshot.Devices))
-				for _, device := range snapshot.Devices {
-					summaries := make([]portal.ProjectionSummary, 0, len(device.Projections))
-					for _, projection := range device.Projections {
-						summaries = append(summaries, portal.ProjectionSummary{
-							Plane:     projection.Plane,
-							NodeCount: len(projection.Nodes),
-							EdgeCount: len(projection.Edges),
-						})
-					}
-					items = append(items, portal.ProjectionDevice{
-						Address:      device.Address,
-						DeviceID:     device.DeviceID,
-						DisplayName:  device.DisplayName,
-						Manufacturer: device.Manufacturer,
-						Projections:  summaries,
-					})
-				}
-				return items
-			},
-			GetProjection: func(address byte, plane string) (portal.ProjectionGraph, bool) {
-				snapshot := builder.FreshSchema()
-				for _, device := range snapshot.Devices {
-					if device.Address != address {
-						continue
-					}
-					for _, projection := range device.Projections {
-						if !strings.EqualFold(projection.Plane, plane) {
-							continue
-						}
-						nodes := make([]portal.ProjectionNode, 0, len(projection.Nodes))
-						for _, node := range projection.Nodes {
-							nodes = append(nodes, portal.ProjectionNode{
-								ID:            node.ID,
-								Path:          node.Path,
-								CanonicalPath: node.CanonicalPath,
-							})
-						}
-						edges := make([]portal.ProjectionEdge, 0, len(projection.Edges))
-						for _, edge := range projection.Edges {
-							edges = append(edges, portal.ProjectionEdge{
-								ID:   edge.ID,
-								From: edge.From,
-								To:   edge.To,
-							})
-						}
-						return portal.ProjectionGraph{
-							Address: address,
-							Plane:   projection.Plane,
-							Nodes:   nodes,
-							Edges:   edges,
-						}, true
-					}
-				}
-				return portal.ProjectionGraph{}, false
-			},
-			ExplorerBus:            gateway.Bus,
-			ExplorerSourceProvider: ebusSourceProvider,
-			// Wire the in-process L7 catalog sub-server (M5_PORTAL).
-			// mcpServer.EbusStandardServer() returns the same instance
-			// RegisterEbusStandardTools installed inside mcp.NewServer;
-			// sharing it between MCP + portal surfaces guarantees both
-			// reach the identical SHA256-pinned embedded catalog. Nil
-			// here would make /api/v1/ebus-standard/* routes 404 in
-			// production (handler.go nil-guard) — see PR #507.
-			EbusStandardServer: mcpServer.EbusStandardServer(),
-		})
-		mux.Handle(portalPath+"/", http.StripPrefix(portalPath, portalHandler))
-		mux.HandleFunc(portalPath, func(w http.ResponseWriter, r *http.Request) {
-			http.Redirect(w, r, portalPath+"/", http.StatusMovedPermanently)
-		})
-	}
-
-	var operatorEndpoint io.Closer
-	if eebusProvider != nil {
-		operatorEndpoint, err = mcpServer.StartEEBusV1OperatorEndpoint(ctx)
-		if err != nil {
-			return nil, nil, fmt.Errorf("start eeBUS operator MCP endpoint: %w", err)
-		}
-	}
-
-	listener, err := net.Listen("tcp", cfg.HTTPAddr)
-	if err != nil {
-		if operatorEndpoint != nil {
-			_ = operatorEndpoint.Close()
-		}
-		return nil, nil, err
-	}
-
-	server := &http.Server{
-		Handler: mux,
-	}
-	if operatorEndpoint != nil {
-		server.RegisterOnShutdown(func() {
-			_ = operatorEndpoint.Close()
-		})
-	}
-
-	go func() {
-		defer func() {
-			if operatorEndpoint != nil {
-				_ = operatorEndpoint.Close()
-			}
-		}()
-		if err := server.Serve(listener); err != nil && err != http.ErrServerClosed {
-			log.Printf("http server error: %v", err)
-		}
-	}()
-
-	var advertiser mdns.Advertiser
-	if cfg.MDNSAdvertise {
-		port := listener.Addr().(*net.TCPAddr).Port
-		advertiser, err = mdns.Advertise(ctx, mdns.Service{
-			Instance: cfg.MDNSInstance,
-			Service:  mdns.ServiceTypeGateway,
-			Port:     port,
-			Text:     gatewayMDNSText(cfg),
-		})
-		if err != nil {
-			_ = server.Close()
-			if operatorEndpoint != nil {
-				_ = operatorEndpoint.Close()
-			}
-			return nil, nil, err
-		}
-	}
-
-	go func() {
-		<-ctx.Done()
-		_ = server.Shutdown(context.Background())
-		if operatorEndpoint != nil {
-			_ = operatorEndpoint.Close()
-		}
-	}()
-
-	return server, advertiser, nil
-}
-
-func auditNumber(value *uint64) string {
-	if value == nil {
-		return "-"
-	}
-	return strconv.FormatUint(*value, 10)
-}
-
-func portalFM5Interpretation(provider graphql.SemanticProvider) graphql.Fm5Interpretation {
-	if provider == nil {
-		return graphql.Fm5Interpretation{}
-	}
-	typed, ok := provider.(graphql.FM5InterpretationProvider)
-	if !ok {
-		return graphql.Fm5Interpretation{}
-	}
-	verdict := typed.FM5Interpretation()
-	if verdict == (graphql.Fm5Interpretation{}) {
-		return verdict
-	}
-	if verdict.Validate() != nil {
-		return graphql.Fm5Interpretation{}
-	}
-	return verdict
-}
-
-func normalizeMountPath(path string, fallback string) string {
-	normalized := strings.TrimSpace(path)
-	if normalized == "" {
-		normalized = fallback
-	}
-	if !strings.HasPrefix(normalized, "/") {
-		normalized = "/" + normalized
-	}
-	if normalized != "/" {
-		normalized = strings.TrimRight(normalized, "/")
-	}
-	if normalized == "/" {
-		return fallback
-	}
-	return normalized
-}
-
-func mapPortalZones(zones []graphql.Zone) []portal.SemanticZone {
-	if len(zones) == 0 {
-		return nil
-	}
-	items := make([]portal.SemanticZone, 0, len(zones))
-	for _, zone := range zones {
-		items = append(items, portal.SemanticZone{
-			ID:   zone.ID,
-			Name: zone.Name,
-			State: portal.SemanticZoneState{
-				CurrentTempC:       cloneFloatPtr(zone.State.CurrentTempC),
-				CurrentHumidityPct: cloneFloatPtr(zone.State.CurrentHumidityPct),
-				HvacAction:         zone.State.HvacAction,
-				SpecialFunction:    zone.State.SpecialFunction,
-				HeatingDemandPct:   cloneFloatPtr(zone.State.HeatingDemandPct),
-				ValvePositionPct:   cloneFloatPtr(zone.State.ValvePositionPct),
-			},
-			Config: portal.SemanticZoneConfig{
-				OperatingMode:              zone.Config.OperatingMode,
-				OperationModeChangeable:    cloneBoolPtr(zone.Config.OperationModeChangeable),
-				SourceLabel:                cloneStringPtr(zone.Config.SourceLabel),
-				Preset:                     zone.Config.Preset,
-				TargetTempC:                cloneFloatPtr(zone.Config.TargetTempC),
-				AllowedModes:               append([]string(nil), zone.Config.AllowedModes...),
-				CircuitType:                zone.Config.CircuitType,
-				AssociatedCircuit:          cloneIntPtr(zone.Config.AssociatedCircuit),
-				RoomTemperatureZoneMapping: cloneIntPtr(zone.Config.RoomTemperatureZoneMapping),
-			},
-		})
-	}
-	return items
-}
-
-func mapPortalDHW(status *graphql.DhwStatus) *portal.SemanticDHW {
-	if status == nil {
-		return nil
-	}
-	return &portal.SemanticDHW{
-		State: portal.SemanticDhwState{
-			CurrentTempC:     cloneFloatPtr(status.State.CurrentTempC),
-			OverrunActive:    cloneBoolPtr(status.State.OverrunActive),
-			SpecialFunction:  status.State.SpecialFunction,
-			HeatingDemandPct: cloneFloatPtr(status.State.HeatingDemandPct),
-		},
-		Config: portal.SemanticDhwConfig{
-			OperatingMode:           status.Config.OperatingMode,
-			OperationModeChangeable: cloneBoolPtr(status.Config.OperationModeChangeable),
-			Preset:                  status.Config.Preset,
-			TargetTempC:             cloneFloatPtr(status.Config.TargetTempC),
-		},
-	}
-}
-
-func mapPortalEnergyTotals(value *graphql.EnergyTotals) *portal.SemanticEnergyTotals {
-	if value == nil {
-		return nil
-	}
-	return &portal.SemanticEnergyTotals{
-		Gas:      mapPortalEnergyChannel(value.Gas),
-		Electric: mapPortalEnergyChannel(value.Electric),
-		Solar:    mapPortalEnergyChannel(value.Solar),
-	}
-}
-
-func mapPortalEnergyChannel(channel graphql.EnergyChannel) portal.SemanticEnergyChannel {
-	return portal.SemanticEnergyChannel{
-		DHW:     mapPortalEnergySeries(channel.DHW),
-		Climate: mapPortalEnergySeries(channel.Climate),
-	}
-}
-
-func mapPortalEnergySeries(series graphql.EnergySeries) portal.SemanticEnergySeries {
-	out := portal.SemanticEnergySeries{
-		Today:     series.Today,
-		TodayMeta: mapPortalEnergyPointMeta(series.TodayMeta),
-	}
-	if len(series.Yearly) > 0 {
-		out.Yearly = append([]float64(nil), series.Yearly...)
-	}
-	if len(series.Monthly) > 0 {
-		out.Monthly = append([]float64(nil), series.Monthly...)
-	}
-	if len(series.YearlyMeta) > 0 {
-		out.YearlyMeta = make([]portal.SemanticEnergyPointMeta, len(series.YearlyMeta))
-		for i, meta := range series.YearlyMeta {
-			out.YearlyMeta[i] = mapPortalEnergyPointMeta(meta)
-		}
-	}
-	if len(series.MonthlyMeta) > 0 {
-		out.MonthlyMeta = make([]portal.SemanticEnergyPointMeta, len(series.MonthlyMeta))
-		for i, meta := range series.MonthlyMeta {
-			out.MonthlyMeta[i] = mapPortalEnergyPointMeta(meta)
-		}
-	}
-	return out
-}
-
-func mapPortalEnergyPointMeta(meta graphql.EnergyPointMeta) portal.SemanticEnergyPointMeta {
-	return portal.SemanticEnergyPointMeta{
-		FreshnessState:  string(meta.FreshnessState),
-		Provenance:      string(meta.Provenance),
-		LastObservedUTC: meta.LastObservedUTC,
-		AgeSeconds:      meta.AgeSeconds,
-		Stale:           meta.Stale,
-	}
-}
-
-func mapPortalBoilerStatus(status *graphql.BoilerStatus) *portal.SemanticBoilerStatus {
-	if status == nil {
-		return nil
-	}
-	return &portal.SemanticBoilerStatus{
-		State: portal.SemanticBoilerState{
-			FlowTemperatureC:         cloneFloatPtr(status.State.FlowTemperatureC),
-			ReturnTemperatureC:       cloneFloatPtr(status.State.ReturnTemperatureC),
-			CentralHeatingPumpActive: cloneBoolPtr(status.State.CentralHeatingPumpActive),
-			WaterPressureBar:         cloneFloatPtr(status.State.WaterPressureBar),
-			ExternalPumpActive:       cloneBoolPtr(status.State.ExternalPumpActive),
-			CirculationPumpActive:    cloneBoolPtr(status.State.CirculationPumpActive),
-			GasValveActive:           cloneBoolPtr(status.State.GasValveActive),
-			FlameActive:              cloneBoolPtr(status.State.FlameActive),
-			DiverterValvePositionPct: cloneFloatPtr(status.State.DiverterValvePositionPct),
-			FanSpeedRpm:              cloneIntPtr(status.State.FanSpeedRpm),
-			TargetFanSpeedRpm:        cloneIntPtr(status.State.TargetFanSpeedRpm),
-			IonisationVoltageUa:      cloneFloatPtr(status.State.IonisationVoltageUa),
-			DhwWaterFlowLpm:          cloneFloatPtr(status.State.DhwWaterFlowLpm),
-			DhwDemandActive:          cloneBoolPtr(status.State.DhwDemandActive),
-			HeatingSwitchActive:      cloneBoolPtr(status.State.HeatingSwitchActive),
-			StorageLoadPumpPct:       cloneFloatPtr(status.State.StorageLoadPumpPct),
-			ModulationPct:            cloneFloatPtr(status.State.ModulationPct),
-			PrimaryCircuitFlowLpm:    cloneFloatPtr(status.State.PrimaryCircuitFlowLpm),
-			FlowTempDesiredC:         cloneFloatPtr(status.State.FlowTempDesiredC),
-			DhwTempDesiredC:          cloneFloatPtr(status.State.DhwTempDesiredC),
-			StateNumber:              cloneIntPtr(status.State.StateNumber),
-			DhwTemperatureC:          cloneFloatPtr(status.State.DhwTemperatureC),
-			DhwTargetTemperatureC:    cloneFloatPtr(status.State.DhwTargetTemperatureC),
-		},
-		Config: portal.SemanticBoilerConfig{
-			DhwOperatingMode: cloneStringPtr(status.Config.DhwOperatingMode),
-			FlowsetHcMaxC:    cloneFloatPtr(status.Config.FlowsetHcMaxC),
-			FlowsetHwcMaxC:   cloneFloatPtr(status.Config.FlowsetHwcMaxC),
-			PartloadHcKW:     cloneFloatPtr(status.Config.PartloadHcKW),
-			PartloadHwcKW:    cloneFloatPtr(status.Config.PartloadHwcKW),
-		},
-		Diagnostics: portal.SemanticBoilerDiagnostics{
-			HeatingStatusRaw:         cloneIntPtr(status.Diagnostics.HeatingStatusRaw),
-			DhwStatusRaw:             cloneIntPtr(status.Diagnostics.DhwStatusRaw),
-			CentralHeatingHours:      cloneFloatPtr(status.Diagnostics.CentralHeatingHours),
-			DhwHours:                 cloneFloatPtr(status.Diagnostics.DhwHours),
-			CentralHeatingStarts:     cloneIntPtr(status.Diagnostics.CentralHeatingStarts),
-			DhwStarts:                cloneIntPtr(status.Diagnostics.DhwStarts),
-			PumpHours:                cloneFloatPtr(status.Diagnostics.PumpHours),
-			FanHours:                 cloneFloatPtr(status.Diagnostics.FanHours),
-			DeactivationsIFC:         cloneIntPtr(status.Diagnostics.DeactivationsIFC),
-			DeactivationsTemplimiter: cloneIntPtr(status.Diagnostics.DeactivationsTemplimiter),
-		},
-	}
-}
-
-func mapPortalSystemStatus(status *graphql.SystemStatus) *portal.SemanticSystemStatus {
-	if status == nil {
-		return nil
-	}
-	return &portal.SemanticSystemStatus{
-		State: portal.SemanticSystemState{
-			SystemOff:                    cloneBoolPtr(status.State.SystemOff),
-			SystemWaterPressure:          cloneFloatPtr(status.State.SystemWaterPressure),
-			SystemFlowTemperature:        cloneFloatPtr(status.State.SystemFlowTemperature),
-			OutdoorTemperature:           cloneFloatPtr(status.State.OutdoorTemperature),
-			OutdoorTemperatureAvg24h:     cloneFloatPtr(status.State.OutdoorTemperatureAvg24h),
-			MaintenanceDue:               cloneBoolPtr(status.State.MaintenanceDue),
-			HwcCylinderTemperatureTop:    cloneFloatPtr(status.State.HwcCylinderTemperatureTop),
-			HwcCylinderTemperatureBottom: cloneFloatPtr(status.State.HwcCylinderTemperatureBottom),
-		},
-		Config: portal.SemanticSystemConfig{
-			AdaptiveHeatingCurve:         cloneBoolPtr(status.Config.AdaptiveHeatingCurve),
-			AlternativePoint:             cloneFloatPtr(status.Config.AlternativePoint),
-			HeatingCircuitBivalencePoint: cloneFloatPtr(status.Config.HeatingCircuitBivalencePoint),
-			DhwBivalencePoint:            cloneFloatPtr(status.Config.DhwBivalencePoint),
-			HcEmergencyTemperature:       cloneFloatPtr(status.Config.HcEmergencyTemperature),
-			HwcMaxFlowTempDesired:        cloneFloatPtr(status.Config.HwcMaxFlowTempDesired),
-			MaxRoomHumidity:              cloneIntPtr(status.Config.MaxRoomHumidity),
-		},
-		Properties: portal.SemanticSystemProperties{
-			SystemScheme:            cloneIntPtr(status.Properties.SystemScheme),
-			ModuleConfigurationVR71: cloneIntPtr(status.Properties.ModuleConfigurationVR71),
-		},
-		GatewayBrand:  cloneStringPtr(status.GatewayBrand),
-		GatewayVendor: cloneStringPtr(status.GatewayVendor),
-	}
-}
-
-func mapPortalCircuits(circuits []graphql.CircuitStatus) []portal.SemanticCircuit {
-	if len(circuits) == 0 {
-		return nil
-	}
-	items := make([]portal.SemanticCircuit, 0, len(circuits))
-	for _, circuit := range circuits {
-		items = append(items, portal.SemanticCircuit{
-			Index:       circuit.Index,
-			CircuitType: circuit.CircuitType,
-			HasMixer:    circuit.HasMixer,
-			State: portal.SemanticCircuitState{
-				PumpActive:       cloneBoolPtr(circuit.State.PumpActive),
-				MixerPositionPct: cloneFloatPtr(circuit.State.MixerPositionPct),
-				FlowTemperatureC: cloneFloatPtr(circuit.State.FlowTemperatureC),
-				FlowSetpointC:    cloneFloatPtr(circuit.State.FlowSetpointC),
-				CalcFlowTempC:    cloneFloatPtr(circuit.State.CalcFlowTempC),
-				CircuitState:     circuit.State.CircuitState,
-				Humidity:         cloneFloatPtr(circuit.State.Humidity),
-				DewPoint:         cloneFloatPtr(circuit.State.DewPoint),
-				PumpHours:        cloneFloatPtr(circuit.State.PumpHours),
-				PumpStarts:       cloneIntPtr(circuit.State.PumpStarts),
-			},
-			Config: portal.SemanticCircuitConfig{
-				HeatingCurve:    cloneFloatPtr(circuit.Config.HeatingCurve),
-				FlowTempMaxC:    cloneFloatPtr(circuit.Config.FlowTempMaxC),
-				FlowTempMinC:    cloneFloatPtr(circuit.Config.FlowTempMinC),
-				SummerLimitC:    cloneFloatPtr(circuit.Config.SummerLimitC),
-				FrostProtC:      cloneFloatPtr(circuit.Config.FrostProtC),
-				RoomTempControl: circuit.Config.RoomTempControl,
-				CoolingEnabled:  cloneBoolPtr(circuit.Config.CoolingEnabled),
-			},
-			ManagingDevice: portal.SemanticManagingDevice{
-				Role:     string(circuit.ManagingDevice.Role),
-				DeviceID: cloneStringPtr(circuit.ManagingDevice.DeviceID),
-				Address:  cloneIntPtr(circuit.ManagingDevice.Address),
-			},
-		})
-	}
-	return items
-}
-
-func mapPortalRadioDevices(devices []graphql.RadioDevice) []portal.SemanticRadioDevice {
-	if len(devices) == 0 {
-		return nil
-	}
-	items := make([]portal.SemanticRadioDevice, 0, len(devices))
-	for _, device := range devices {
-		items = append(items, portal.SemanticRadioDevice{
-			Group:                device.Group,
-			Instance:             device.Instance,
-			SlotMode:             device.SlotMode,
-			DeviceConnected:      cloneBoolPtr(device.DeviceConnected),
-			DeviceClassAddress:   cloneIntPtr(device.DeviceClassAddress),
-			DeviceModel:          device.DeviceModel,
-			FirmwareVersion:      cloneStringPtr(device.FirmwareVersion),
-			HardwareIdentifier:   cloneIntPtr(device.HardwareIdentifier),
-			RemoteControlAddress: cloneIntPtr(device.RemoteControlAddress),
-			DevicePaired:         cloneBoolPtr(device.DevicePaired),
-			ReceptionStrength:    cloneIntPtr(device.ReceptionStrength),
-			ZoneAssignment:       cloneIntPtr(device.ZoneAssignment),
-			RoomTemperatureC:     cloneFloatPtr(device.RoomTemperatureC),
-			RoomHumidityPct:      cloneFloatPtr(device.RoomHumidityPct),
-		})
-	}
-	return items
-}
-
-func mapPortalSolarStatus(status *graphql.SolarStatus) *portal.SemanticSolarStatus {
-	if status == nil {
-		return nil
-	}
-	return &portal.SemanticSolarStatus{
-		CollectorTemperatureC: cloneFloatPtr(status.CollectorTemperatureC),
-		ReturnTemperatureC:    cloneFloatPtr(status.ReturnTemperatureC),
-		PumpActive:            cloneBoolPtr(status.PumpActive),
-		CurrentYield:          cloneFloatPtr(status.CurrentYield),
-		PumpHours:             cloneFloatPtr(status.PumpHours),
-		SolarEnabled:          cloneBoolPtr(status.SolarEnabled),
-		FunctionMode:          cloneBoolPtr(status.FunctionMode),
-	}
-}
-
-func mapPortalCylinders(cylinders []graphql.CylinderStatus) []portal.SemanticCylinder {
-	if len(cylinders) == 0 {
-		return nil
-	}
-	items := make([]portal.SemanticCylinder, 0, len(cylinders))
-	for _, cylinder := range cylinders {
-		items = append(items, portal.SemanticCylinder{
-			Index:             cylinder.Index,
-			TemperatureC:      cloneFloatPtr(cylinder.TemperatureC),
-			MaxSetpointC:      cloneFloatPtr(cylinder.MaxSetpointC),
-			ChargeHysteresisC: cloneFloatPtr(cylinder.ChargeHysteresisC),
-			ChargeOffsetC:     cloneFloatPtr(cylinder.ChargeOffsetC),
-		})
-	}
-	return items
-}
-
-func mapPortalAdapterInfo(info *graphql.AdapterHardwareInfo) *portal.SemanticAdapterInfo {
-	if info == nil {
-		return nil
-	}
-	result := &portal.SemanticAdapterInfo{
-		FirmwareVersion:    info.FirmwareVersion,
-		FirmwareChecksum:   info.FirmwareChecksum,
-		BootloaderVersion:  info.BootloaderVersion,
-		BootloaderChecksum: info.BootloaderChecksum,
-		HardwareID:         info.HardwareID,
-		HardwareConfig:     info.HardwareConfig,
-		Features:           info.Features,
-		Jumpers:            info.Jumpers,
-		IsWiFi:             info.IsWiFi,
-		IsEthernet:         info.IsEthernet,
-		VersionResponseLen: info.VersionResponseLen,
-		InfoSupported:      info.InfoSupported,
-		TemperatureC:       cloneFloatPtr(info.TemperatureC),
-		SupplyVoltageMV:    cloneIntPtr(info.SupplyVoltageMV),
-		BusVoltageMaxDV:    cloneIntPtr(info.BusVoltageMaxDV),
-		BusVoltageMinDV:    cloneIntPtr(info.BusVoltageMinDV),
-		ResetCause:         cloneStringPtr(info.ResetCause),
-		WiFiRSSIDBm:        cloneIntPtr(info.WiFiRSSIDBm),
-	}
-	if info.JumperFlags != nil {
-		result.JumperFlags = make([]string, len(info.JumperFlags))
-		copy(result.JumperFlags, info.JumperFlags)
-	}
-	if info.ResetCauseCode != nil {
-		code := int(*info.ResetCauseCode)
-		result.ResetCauseCode = &code
-	}
-	if info.RestartCount != nil {
-		count := int(*info.RestartCount)
-		result.RestartCount = &count
-	}
-	if info.LastIdentityQuery != nil {
-		result.LastIdentityQuery = info.LastIdentityQuery.Format(time.RFC3339)
-	}
-	if info.LastTelemetryQuery != nil {
-		result.LastTelemetryQuery = info.LastTelemetryQuery.Format(time.RFC3339)
-	}
-	return result
-}
-
-// buildResponderCapabilityProvider composes the `meta.capabilities.responder`
-// signal for the active transport per decision doc @ 567a6798 §4.2 + §5.
-//
-// Mapping (v1.1, locked):
-//   - ENH / ENS → active.scope = "partial", surfaces = FF_03..FF_06,
-//     refusal = nil. transports[] reports the same row as "supported"
-//     and the ebusd-tcp row as perpetually "blocked".
-//   - ebusd-tcp → active.scope = "none", surfaces = [],
-//     refusal.code = "command_bridge_no_companion_listen". Consumers
-//     apply fail-closed per §4.3 rule 4.
-//   - Any other transport (udp-plain, tcp-plain, adapter-direct, empty,
-//     or unrecognised string) → returns nil. The caller omits the
-//     capability entirely and consumers fall back to §4.3 rule 1
-//     (absence → scope=none, fail-closed). This is the only way to
-//     preserve invariant I1 (exactly three rows at v1.1) AND I2
-//     (active.transport MUST appear in transports[]) simultaneously,
-//     because emitting a non-canonical active.transport literal would
-//     violate I2 and widening transports[] with a fourth "unknown" row
-//     would violate I1.
-//
-// The raw cfg.TransportConfig.Protocol is first canonicalised via
-// ebusgateway.canonicalTransportProtocol (which handles the "ebusd"
-// alias → "ebusd-tcp" and lowercases/trims whitespace). This ensures
-// downstream envelope assertions see the canonical enum literal in
-// every surface, matching the three fixed transports[] rows byte-for-
-// byte.
-//
-// Invariants I1/I7 are enforced by always emitting exactly three
-// transport rows (ENH, ENS, ebusd-tcp) in a fixed order. I2/I3 are
-// enforced by deriving active.transport from the same canonical enum
-// literals the rows use.
-//
-// Runtime-transport authority (Codex P1 on PR #509): the canonical
-// protocol enum alone is NOT sufficient to decide active.scope. The
-// adapter-direct URI mode (--address=adapter-direct://...) keeps
-// TransportConfig.Protocol as the default "enh" string while the live
-// transport instance is actually the adapter-direct mux's active path.
-// That path implements transport.RawTransport but does NOT implement
-// transport.ResponderTransport — it has no SendResponderBytes primitive.
-// Reporting active.scope="partial" purely from the config string would
-// over-advertise responder support for FF_03..FF_06 surfaces the gateway
-// cannot actually emit. Accordingly the provider type-asserts the live
-// transport instance against ebusgoTransport.ResponderTransport; when
-// the canonical protocol is ENH/ENS but the instance does NOT satisfy
-// the interface, we downgrade to scope="none" with
-// refusal.code="transport_mux_bypass" AND rewrite both ENH and ENS
-// rows in transports[] to state="blocked", scope="none",
-// reason="transport_mux_bypass" so invariant I3 (active.scope ==
-// matching row.scope) holds — the mux is a shared runtime wrapper
-// above either upstream, so switching the canonical protocol would
-// not restore responder emission. ebusd-tcp path is unchanged —
-// it is forbidden from responder emission per M4b2 §3 regardless of
-// what underlying transport type ebusd is wrapped by.
-//
-// actualTransport is the live instance returned by the bootstrap
-// transport factory. A nil value means the caller has not yet wired a
-// transport (legacy callers, some unit-test paths); in that case the
-// provider falls back to protocol-only inference to preserve previous
-// behaviour on paths that never exercise the adapter-direct mux.
-func buildResponderCapabilityProvider(cfg ebusgateway.Config, actualTransport ebusgoTransport.RawTransport) ebus_standard.ResponderCapabilityProvider {
-	surfacesFF := []string{"FF_03", "FF_04", "FF_05", "FF_06"}
-	// transports[] is static at v1.1 — same three rows on every gateway
-	// regardless of deployment (I1 locks the count at exactly three).
-	transports := []ebus_standard.TransportRow{
-		{Transport: "ENH", State: "supported", Scope: "partial", Surfaces: surfacesFF, Reason: ""},
-		{Transport: "ENS", State: "supported", Scope: "partial", Surfaces: surfacesFF, Reason: ""},
-		{Transport: "ebusd-tcp", State: "blocked", Scope: "none", Surfaces: []string{}, Reason: "command_bridge_no_companion_listen"},
-	}
-
-	// Runtime transport authority: a non-nil actualTransport that does
-	// NOT satisfy ResponderTransport means the live bus path cannot
-	// actually emit responder bytes (the adapter-direct mux is the
-	// concrete case today). A nil actualTransport means bootstrap is
-	// still pre-wiring (e.g. unit tests that only construct Config); in
-	// that case fall back to protocol-only inference so legacy callers
-	// keep their behaviour.
-	transportKnown := actualTransport != nil
-	_, responderCapable := actualTransport.(ebusgoTransport.ResponderTransport)
-	muxBypass := transportKnown && !responderCapable
-	muxBypassRefusal := &ebus_standard.ActiveRefusal{
-		Code:   "transport_mux_bypass",
-		Reason: "runtime transport does not satisfy ResponderTransport (e.g. adapter-direct mux)",
-	}
-
-	// Invariant I3 (decision doc §4.4): active.scope MUST equal
-	// transports[x].scope where x.transport == active.transport. When the
-	// runtime mux bypass is in effect we therefore also rewrite the
-	// matching transports[] row(s) — otherwise a consumer joining
-	// active.transport → transports[] would see contradictory metadata
-	// (active.scope=none but row.scope=partial).
-	//
-	// Interpretation A (shared-runtime downgrade): the adapter-direct mux
-	// is a single wrapper that sits above whichever ENH/ENS upstream the
-	// operator configured. The mux instance itself does not satisfy
-	// ResponderTransport, so switching the canonical protocol from ENH to
-	// ENS (or vice versa) under the same mux would not restore responder
-	// emission — the bypass is shared. Accordingly, BOTH the ENH and ENS
-	// rows downgrade to state=blocked, scope=none, reason="transport_mux_bypass".
-	// The ebusd-tcp row is left untouched (it is always blocked with its
-	// own reason "command_bridge_no_companion_listen" per §4.2/§3).
-	// Invariants preserved: I1 (still exactly 3 rows), I2 (active.transport
-	// still appears verbatim), I3 (active.scope == matching row.scope),
-	// I5 (state=blocked ⇒ reason != ""), I7 (row order ENH,ENS,ebusd-tcp
-	// unchanged).
-	if muxBypass {
-		for i := range transports {
-			row := &transports[i]
-			if row.Transport == "ENH" || row.Transport == "ENS" {
-				row.State = "blocked"
-				row.Scope = "none"
-				row.Surfaces = []string{}
-				row.Reason = "transport_mux_bypass"
-			}
-		}
-	}
-
-	var active ebus_standard.ActiveResponder
-	// Canonicalise raw TransportProtocol (handles "ebusd" alias, case,
-	// whitespace) into one of the enum constants so active.transport
-	// literally equals the transports[] row label (invariant I2).
-	switch ebusgateway.CanonicalTransportProtocol(cfg.TransportConfig.Protocol) {
-	case ebusgateway.TransportENH:
-		if muxBypass {
-			active = ebus_standard.ActiveResponder{Transport: "ENH", Scope: "none", Surfaces: []string{}, Refusal: muxBypassRefusal}
-		} else {
-			active = ebus_standard.ActiveResponder{Transport: "ENH", Scope: "partial", Surfaces: surfacesFF}
-		}
-	case ebusgateway.TransportENS:
-		if muxBypass {
-			active = ebus_standard.ActiveResponder{Transport: "ENS", Scope: "none", Surfaces: []string{}, Refusal: muxBypassRefusal}
-		} else {
-			active = ebus_standard.ActiveResponder{Transport: "ENS", Scope: "partial", Surfaces: surfacesFF}
-		}
-	case ebusgateway.TransportEbusdTCP:
-		active = ebus_standard.ActiveResponder{
-			Transport: "ebusd-tcp",
-			Scope:     "none",
-			Surfaces:  []string{},
-			Refusal:   &ebus_standard.ActiveRefusal{Code: "command_bridge_no_companion_listen", Reason: "ebusd command bridge does not expose a responder-role emission primitive"},
-		}
-	default:
-		// Non-enumerated transports (udp-plain, tcp-plain,
-		// adapter-direct, empty, or entirely unknown). Fail-closed:
-		// return nil so the envelope omits meta.capabilities.responder
-		// entirely (§4.3 rule 1). Emitting the raw string would
-		// violate I2; adding a fourth transports[] row would violate
-		// I1. Absence is the only invariant-preserving outcome.
-		return nil
-	}
-
-	cap := ebus_standard.ResponderCapability{
-		Version:    "v1",
-		Active:     active,
-		Transports: transports,
-	}
-	return func() ebus_standard.ResponderCapability { return cap }
-}
-
-// applyStaticSeedTable plants the productids static seed entries into
-// the registry. Each seed entry contributes one DeviceInfo per address
-// with full Vaillant identity (Manufacturer + DeviceID), allowing the
-// registry's identity-merge contract to collapse canonical-pair faces
-// into a single entry. SerialNumber and version fields are
-// intentionally empty — they will be populated by subsequent active
-// enrichment (P5 follow-up) or remain empty for seed-only addresses
-// (e.g. NETX3 broadcast face 0x04 which does not respond to active
-// probes).
-//
-// Phase post-C P3 (live validation 2026-05-08): NETX3's 0x04 face was
-// absent from the registry entirely because broadcast-source frames
-// never carry an ACKCorrelation that would feed the inserter. Static
-// seed bypasses that gate at startup.
-//
-// P3.5 (Codex P2 follow-up, ebusreg PR #137): switches the per-address
-// call from registry.Register (which stamps the AddressSlot with
-// DiscoverySourceActiveConfirmed/VerificationStateIdentityConfirmed —
-// wrong observability label for a pre-known seed entry) to
-// registry.RegisterStaticSeed (which stamps DiscoverySourceStaticSeed
-// /VerificationStateCandidate). Operators reading
-// `ebus.v1.registry.devices.list` or the address-table snapshot via
-// MCP/JSON now correctly see seeded addresses labelled
-// `discovery_source: "static_seed"`, `verification: "candidate"`
-// instead of pretending the gateway actively confirmed them at boot.
-//
-// Role mapping happens at this seam (gateway), not in productids or
-// registry: productids.SeedAddressEntry.Role is a free-form string
-// (`"initiator"` / `"target"`); registry.SlotRole is a typed enum
-// (SlotRoleMaster / SlotRoleSlave / SlotRoleUnknown). Unknown role
-// strings fall through to SlotRoleUnknown — registry's monotonic Role
-// guard then leaves Role empty until passive observation or active
-// scan fills it in.
-func applyStaticSeedTable(reg *registry.DeviceRegistry) {
-	seeds := productids.LoadSeedTable(true)
-	if len(seeds) == 0 {
-		return
-	}
-	now := time.Now()
-	count := 0
-	for _, seed := range seeds {
-		for _, addr := range seed.Addresses {
-			role := registry.SlotRoleUnknown
-			switch addr.Role {
-			case "initiator":
-				role = registry.SlotRoleMaster
-			case "target":
-				role = registry.SlotRoleSlave
-			}
-			info := registry.DeviceInfo{
-				Address:      addr.Addr,
-				Manufacturer: seed.Manufacturer,
-				DeviceID:     seed.DeviceID,
-			}
-			reg.RegisterStaticSeed(info, role, now)
-			count++
-		}
-	}
-	log.Printf("static seed table: planted %d address(es) across %d device(s) at startup (source=productids.LoadSeedTable, label=static_seed/candidate)", count, len(seeds))
-}
-
-// initRuntimeStateManager constructs and starts the runtime-state Manager,
-// wiring AD08 eager-persist for the instance-guid CLI flag (when present)
-// and exposing the loaded state for hint extraction.
-//
-// On Manager.Load failure (missing / corrupt file), Manager.Load returns an
-// empty state and the gateway continues without a hint — the cache is a
-// best-effort optimisation, not a startup requirement (AD11 + M2 spec).
-func initRuntimeStateManager(ctx context.Context, cfg ebusgateway.Config, buildInfo gatewayBuildInfo) (*runtimestate.Manager, *runtimestate.State) {
-	mgr := runtimestate.New(runtimestate.Options{
-		Path:         cfg.RuntimeStatePath,
-		GatewayBuild: gatewayBuildString(buildInfo),
-		AddonVersion: buildInfo.ReleaseVersion,
-	})
-	state, err := mgr.Load(ctx)
-	if err != nil {
-		log.Printf("runtime_state: load returned error (continuing with empty state): %v", err)
-		state = &runtimestate.State{}
-	}
-
-	// AD08 eager persist: when -instance-guid is supplied, durably write
-	// meta.{schema_version, instance_guid, written_at} within ~1 s so the
-	// crash-before-first-periodic-persist window is closed. Provenance per
-	// AD27.
-	if cfg.InstanceGUID != "" {
-		source := identitySourceFromCfg(cfg.InstanceGUIDSource)
-		if perr := mgr.EagerPersistInstanceGUID(ctx, cfg.InstanceGUID, source); perr != nil {
-			log.Printf("runtime_state: eager-persist instance_guid failed (continuing): %v", perr)
-		}
-	} else if perr := mgr.Flush(ctx); perr != nil {
-		log.Printf("runtime_state: startup metadata persist failed (continuing): %v", perr)
-	}
-
-	if serr := mgr.Start(ctx); serr != nil {
-		log.Printf("runtime_state: start failed (continuing without periodic persister): %v", serr)
-	}
-
-	return mgr, state
-}
-
-func projectGatewayReadiness(ebusProxyReadiness func() string, snapshot eebusRuntimeLifecycleSnapshot) portal.RuntimeReadiness {
-	eebusReadiness := eebusReadinessForLifecycle(snapshot)
-	proxyReadiness := ebusProxyReadinessDisabled
-	if ebusProxyReadiness != nil {
-		proxyReadiness = ebusProxyReadiness()
-	}
-	return portal.RuntimeReadiness{
-		ProcessReadiness:    string(eebusReadiness.ProcessReadiness),
-		HTTPReadiness:       "READY",
-		ProxyReadiness:      proxyReadiness,
-		EEBusReadiness:      string(eebusReadiness.EEBusReadiness),
-		EEBusDegradedReason: string(eebusReadiness.EEBusDegradedReason),
-	}
-}
-
-// identitySourceFromCfg maps the CLI -instance-guid-source flag value to the
-// runtimestate.IdentitySource enum, with AD27 deprecation log when the flag
-// is absent. The well-formed value set is enforced here; unknown values fall
-// back to "cli-override" with a warning rather than failing startup.
-func identitySourceFromCfg(raw string) runtimestate.IdentitySource {
-	switch raw {
-	case string(runtimestate.IdentitySourceRuntimeState):
-		return runtimestate.IdentitySourceRuntimeState
-	case string(runtimestate.IdentitySourceLegacyMigrated):
-		return runtimestate.IdentitySourceLegacyMigrated
-	case string(runtimestate.IdentitySourceGenerated):
-		return runtimestate.IdentitySourceGenerated
-	case string(runtimestate.IdentitySourceCLIOverride):
-		return runtimestate.IdentitySourceCLIOverride
-	case "":
-		log.Print("runtime_state: -instance-guid-source absent; defaulting to cli-override (deprecated; pass -instance-guid-source explicitly)")
-		return runtimestate.IdentitySourceCLIOverride
-	default:
-		log.Printf("runtime_state: -instance-guid-source=%q is not a recognised AD27 value; treating as cli-override", raw)
-		return runtimestate.IdentitySourceCLIOverride
-	}
 }
